Implement TSEM control plane.
The fs.c file contains the implementation of the TSEM control
plane that is in the form of a pseudo-filesystem mounted on the
following directory:

/sys/fs/tsem

The following file documents the interface provided by the
control plane:

Documentation/ABI/testing/tsemfs

The pseudo-files act on the modeling context of the process that
is acting on the file.  For example, reading the 'id'
pseudo-file, returns the modeling domain identifier that the
process is running in.

The ExternalTMA directory is used to segreate the pseudo-files
that are created in order to surface security event descriptions
to an external trust orchestrator.  The files in this directory
appear as the numeric value of the modeling domain they were
created for.

The 'control' pseudo-file is the only writable file in the plane
and is used to control the TSEM implementation.  The most
important and primary roles are to create namespaces and set the
trust status of a process modeled by an external TMA.

Signed-off-by: Greg Wettstein <greg@enjellic.com>