Implement security event mapping.
The map.c file is responsible for implenting the description of a
security event into a security event state point.  The following
documentation file provided as a part of the TSEM implementation
contains a description of this mapping process:

Documentation/admin-guide/LSM/tsem.rst

The mapping process takes a security event description, that was
described for the introduction of the event.c file, and uses that
to drive the mapping process.  The allocation and mapping of the
event is unified through the tsem_map_event() function provided
in this file.

The function for a security event state point mapping is as follows:

Sp = SHA256(SHA256(EVENT_ID) || TASK_ID || SHA256(COE) || SHA256(CELL))

This function is fully described in the previously noted
documentation file.

The TASK_ID is the security state point for the
bprm_creds_for_exec security event hook.  It is generated by the
tsem_map_task() function that is implemented in this file.

The TASK_ID mapping function uses the same functional expression
as the security state point mapping but substitutes a TASK_ID
that consists of 32 null bytes.

Since a generically modeled event does not have a format CELL
definition this file contains a default CELL digest value that is
used in the generation for a security state point for these
events.  A generic value is currently supplied with a roadmap
that would allow trust orchestrators to set this value in order
to avoid a kernel ABI dependency on this value.

Signed-off-by: Greg Wettstein <greg@enjellic.com>