Configure tsem to be the first LSM in the initialization and call order.
The first release of the patch set drew a question from Casey
Schaufler as to whether or not tsem should be the last LSM in the
call order.  He indicated that we should check with the the group
that is implementing the BPF LSM to find out if they objected to
having tsem come after them.

This appears to be secondary to the fact that the BPF
implementation may be modifying how the LSM stack actually
operates.

I had composed an e-mail reply in response to the patch review
questioning if there was a recommendation as to where TSEM should
be positioned.  I noted that, given that TSEM could be used to
conduct activities such as validating the LSM security
attributes, that it would make sense to have TSEM be first in the
call order.

To date I have not received a reply to my question so I moved
TSEM to be first in the call and initialization order until when
and if additional direction is supplied.

TSEM appears to function, after testing, properly in this
configuration.