Convert TSEM control plane to securityfs.
One of the review comments from the the first release cycle of
the TSEM driver was from Greg Kroah-Hartman who objected to
having an independent filesystem implementation for the TSEM
control plane.

I had indicated that the lead of SeLinux and SMACK were followed
who had implemented their control plane as an independent
filesystem whose mount point was in /sys/fs.

We had been operating under the assumption that the securityfs
filesystem was problematic with respect to supporting pollable
files.  After receiving the objection to the independent
filesystem implementation a brief review of the code suggested
that there was probably not an inherent barrier to using pollable
files in securityfs.

This commit drops the infrastructure that was needed to implement
an independent filesystem and converts the control plane to using a
filesystem heirarchy that is based in the following directory:

/sys/kernel/security/tsem

Where /sys/kernel/security is the mount point for the securityfs
filesystem.

These changes allowed the tsem_fs_remove_external() function to
be dropped, since its functionality can be replaced with a call
to the securityfs_remove() function.  The magic number for the
tsemfs filesystem can be dropped which reduces the external
dependencies by one file.