Update handling of modeling domain reference counting and release.
I had received an e-mail from Hillf Danton in response to the
initial posting of the TSEM patches.  Hillf noted in that e-mail
that there was an error in the tsem_ns_put() function that would
result in the release of the modeling context not being
conducted.

Hillf had provided the following explanation:

Given ctx->kref is 2, in the below scenario

	cpu 0		cpu 2
	---		---
	ctx->kref > 1
			ctx->kref > 1
			kref_put
	kref_put

no work will be scheduled, so it makes sense to move scheduling work to
tsem_ns_free() by making tsem_ns_put() only a wrapper of kref_put(), if
ctx has to be released in workqueue.

void tsem_ns_put(struct tsem_TMA_context *ctx)
{
	kref_put(&ctx->kref, tsem_ns_free);
}

Hillf had also noted that, given the availability of the
system_unbound_wq and system_wq, there was probably no reason
to create a TSEM specific workqueue for releasing the modeling
context.

Based on these recommendations the modeling context release
handling was significantly modified and simplified.

The 'release_wq' that was specific to TSEM was dropped in favor
of using the 'system_wq'.  Since one of system supplied work
queues is now being used the tsem_ns_init() function could also
be dropped.

The tsem_ns_put() function was modified to be simply a wrapper
around a call to the kref_put() function that specified that the
tsem_ns_free() function should be called on the release of the
last reference to the domain.

This code review ended up indicating that the tsem_ns_free()
function only needed to be referenced within the namespace.c
file, so it was converted to a statically scoped function.

The new ns_free() function now schedules the work that was
previously done in the tsem_ns_free() function for execution in
workqueue context.

Further review of the implementation also indicated that there
was only a single call to the tsem_ns_get() function, so its
implementation was replaced with a call to the kref_get()
function against the modeling context reference structure.

Documentation for the functionality in namespace.c was also
updated.