Convert trajectory file format to JSON.
In the first round of TSEM patch reviews Greg Kroah-Hartman
raised an issue with the format that was being used to output
security event descriptions.  The consensus appears to be that
encoding the security event descriptions in JSON would be
considered appropriate.

This commit begins the process of this conversion by converting
the standard and forensic execution tranjectories to be output in
JSON format.

The following two globally visible functions were added to assist
in this conversion process:

tsem_fs_show_field()

tsem_fs_show_key()

The first function is used to output a label that will be used to
name a security event description field.  This label is
equivalent to the label that was used to identify a bracketed
field in the previous output format, ie:

event{process=comm, filename=fname, ...}

The second function is used to output a key/value pair within the
event field.

The JSON format being used codes all of the key/value pairs as
strings in order to simplify the output code and allow the trust
orchestrators to determine how the values should be interpreted.

Each output line is an independent JSON encoding of an event and
is thus enclosed in a set of brackets to designate it as a
structure encoding.