Convert security event export to JSON formatting.
Consistent with the security trajectories being output from the
kernel based TMA, convert the export of security event
descriptions to JSON encoding.

Since the external export of a security event description is
identical to the export of a security event description from a
TMA, the external export was converted to use the same
seq_file implementation that the internal TMA uses.  This
minimizes the need to keep two largely identical implementations
in synchronization.

The following external function was implemented to access the
security description export functionality:

tsem_fs_show_trajectory

The PID of the process being modeled is now included in the
event{} description field rather than as a separate field.  In
order to have the same export code properly handle this the
internal TMA sets the PID that is stored in the event description
structure to a value of zero.  The security event export code
only include the pid= key/value pair in the export if the PID
value in the event description structure is non-zero.

In order to provide a more concise description the enumeration
type that is used to specify the type of the export was converted
from 'export_events_type' to 'export_type'.