Implement process authentication.
One of the concerns voiced by Paul Moore, the Linux security
sub-system maintainer, in the initial patch series was the safety
of used a PID as a communications medium between a trust
orchestrator and the kernel.  He noted that PID values were
inherently racy, opening the door for an adversary to substitute
a process of their choosing to be acted upon by a trust
orchestrator.

I had provided an analysis of why I believed that the current
architecture was protected by such an attack by virtue of
regulating what processes can send kill signals to another
process.  Paul brought up the issue of the OOM killer being used
to kill a process in a modeling domain that would allow a
process substitution attack.

The solution to this problem that was elected was to implement
the requirement that a process authenticate itself before a
control call from a trust orchestrator is allowed to proceed.o

The model implemented was to have a trust orchestrator provide,
as part of the external namespace control call, a 256 bit key
value that authenticates the trust orchestrator to a process.
The process that is being delegated to an alternative modeling
process randomly selects a 256 bit key value as part of the
namespace setup.

An authentication key for the trust orchestrator is generated
with the following function:

TMAkey = SHA256(task_key || TMA_authentication_key)

Where the TMA_authentication_key is the binary representation of
the key that is supplied in ASCII format as an argument to the
trusted/untrusted control command.

The randomly generated task_key is placed in the TSEM specific
task control structure of the process being placed in a new
modeling domain.

The TMAkey value is placed in the TSEM specific task control
structure of the parent process of the process that is initiating
a change to a new modeling namespace.

The control_COE() function re-computes the TMAkey value at when a
trust control call is issued by the trust orchestrator based on
the key value supplied as an argument to the control call and the
task_key in the process whose trust status is being set.  If the
generated value does not match the value in the process issueing
the control call a system error alert is generated and no action
on the process is taken.

A list of the currently generated TMAkey values is maintained for
all active security namespaces in order to prevent a TMAkey value
from being accidentally used by another process.

This architecture requires that an adversary be able to predict
the 256 bit random key that was generated at the time the
namespace was created.  An additional and alternative proposal
would involve using the event id number to purturb the TMAkey
value on an event by event basis.  This would require a hash
call on each process control event with its attendant performance
cost.

In order to support this new architecture the following new
function with global scope was added:

tsem_ns_event_key()

Which localizes the generation of the TMAkey into a common
function.

The signature for the tsem_ns_create() function was modified in
order to pass the ASCII representation of the TMA authentication
key to the namespace creation process.

Since the tsen_ns_event_key() function is called with a hash
function context structure, files that were including the
crypto/hash.h include file dropped that include in favor of
having it in the tsem.h file.

Locking was added to the tsem_ns_create() function in order to
close a possible race condition in the allocation of a
context_id.  The mutex used for locking is also used to protect
the linked list that holds the TMAkey's that have been
allocated.

Significant cleanups were also implemented in argument handling
in the write_control() function.