Fully disable TSEM when it is isn't enabled as an LSM.
A regression was noted when TSEM was removed, via the lsm=
command-line option.  A kernel in this configuration would
generate a segmentation fault as it was booting when the
tsem:set_ready() function this was called.

This caused by the tsem:set_ready() function assuming that the
LSM was properly initialized when it was not.

In order to address this the following file scoped boolean
variable was added:

tsem_available

This variable is set, upon proper completion of the tsem_init()
function, to true.

The tsem:set_ready() function checks this variable and if it is
not set to true the call to the function terminates without
conducting the final actions needed to enable the LSM.