Output file description in non-anonymous mmap_file event.
Previously the description of a file that is being memory mapped
was being output as a separate event file.  This ended up having
the mmap_file event being a special case from all of the other
events.

If a file backed mapping is detected the file description is now
output as a component of the mmap_file event description which
allows all of the parameters of an event to be extracted by
simply extracting for the security event type.