Output file_open as a properly named event.
Previously the CELL description for the file_open event was
output as a file{} field description.  This, along with the
previously address mmap_file issue, were the only events that
were not strictly self-defining by virtue of the type= key/value
in the event{} field.

Update the output of a file description field to that the file
description is in a file_open{} event description structure for a
file_open event description.