Allow configuration of root model trust register.
One of the feedbacks from the initial round of review was from
Paul Moore, the Linux security sub-system maintainer, regarding
the selection of the TPM Platform Configuration Register that the
security coefficients for events in the root modeling domain are
extended into.  Paul indicated that he felt this value should be
configurable.

The TSEM configuration was updated to allow selection of the
platform configuration register that is used for the root domain.
The default value is specified as 11 with an allowed range of 8
through 14.