Update handling of COE effective capabilities.
The previously implementation of the code that adds the effective
capability of a context of execution was using the
security_capget() function to retreive the effective, inherited
and permitted capability masks.  There was no real reson to use
that function so the mask is now set by using the current_cred()
function to read the effective capabilities.

In order to increase the clarity of the code the union in the
tsem_COE structure that holds the capability was changed from
'capability' to 'capeff'.

The effective capability tag for the security event description
was changed from 'cap' to 'capeff' to maintain this consistency.