Implement selection of digest function.
This commit implements the ability to select the hash function
that will be used for a modeling domain/namespace.

The specification of the hash function immediately follows the
'internal' or 'external' keyword that selects the location of
where the Trusted Modeling Agent for the domain should be
implemented.

The digest function keyword will be the digesdt implementation name
that is used in the Linux kernel.  If the digest name does not
match an implemented digest in the kernel the control message
write fails.

The character buffer in the tsem_TMA_context structure is
dynamically allocated so the wq_put function was modified to
release the memory allocated for the character buffer as part of
the namespace/domain retirement procedure.

There are currently correctness issues with respect to the digest
value for a faile that is cached in the tsem_inode structure.
The next update to this encounter will address that issue.