Remove conditionalization of network hooks.
An artifact of an early version of the implementation was that a
conditional had been used to eliminate inclusion of the security
hooks that are implemented when the CONFIG_NETWORK_SECURITY
configuration feature is in effect.

The TSEM configuration 'selects' this option so it will always be
present, a fact that eliminates the need to conditionalize away
the security hooks.