Implement standardized zero digest handling.
With the implementation of fully selectable digest functions for
a modeling domain there is a need to address the fact that the
handling of a file of zero size was being addressed by zopying
the 'zero message' SHA256 digest value into the digest value for
the inode.  This had been handled in the security/tsem/event.c
file by converting the ASCII hexadecimal representation of the
SHA256 zero message digest into its binary representation in the
inode digest cache.

The modeling domain/namespace creation functionality was enhanced
to compute the 'zero message' value for whatever hash function
has been selected for the domain.  Some of the hash functions in
the Linux shash implementation have values defined for the 'zero
message' digest but coverage is not complete.  Explicitly
computing the zero message digest value simplifies the
implementation and carries on with the design concept of the only
information that is needed is the name of a supported hash
function.

In order to simplify the tsem_ns_create() function, the
implementation of the digest recognition and parameterization was
moved into the security/tsem/namespace.c:configure_digest
function.

A byte array named zero_digest was added to the tsem_TMA_context
structure to hold the zero message digest value for the hash
function being used to implement the modeling domain.  The
security/tsem/event.c:add_file_digest() function was modified to
use the value held in the modeling context structure for either a
file with zero size or an inode for which pseudonym processing
has been requested.

The security/tsem/model.c:tsem_model_has_peudonym() function was
modified to remove the terminal argument that was pointing to
the byte array that was to be filled in with the pseudonym value.
Since the implementation is now standardizing on a pseudonym
value of the zero message digest size all that is needed is a
determination of whether or not the inode has been declared to be
a pseudonym.