Implement configuration of root domain digest.
With support for the ability of subordinate modeling domains to
use multiple digests, support was added for setting the
compression digest to be used for the root modeling domain.

The default digest function is selected with the following kernel
command-line option:

tsem_digest=DIGESTNAME

Where DIGESTNAME is replaced with a hash function that is
supported by the kernel.

The security/tsem/tsem.c:configure_root_digest() function was
added to encapsulate the functionality needed to configure the
root hash.  If the tsem_digest command-line option has not been
used to configure a hash function the default of sha256 is used.

The size of the digest is placed in the root modeling context
along with the zero message digest.  If a digest function
different than sha256 was selected memory is allocated for the
string and placed in the root context.