Properly compute the address of an accepted UNIX socket.
In the case of a socket accept for a UNIX domain socket the
address digest was being computed from the digest field of the
inode.  Subsequent review of the code suggested that the code may
have been computing the hash of a zero filled field and in any
case it was not clear the behavior was correct.

The implementation was updated, so that in the case of a UNIX
domain socket accept, a pointer to the UNIX domain socket was
passed to the event collection infrastructure.  This pointer
replaces the tsem_inode structure pointer that was previously
being passed.

A new function, security/tsem/event.c:get_socket_accept(), was
added to compute the digest value of the socket address.   This
function does nothing if the socket is IPV4 or IPV6 based.  For a
UNIX domain socket the digest of the socket address is computed.

For other socket types, the zero message digest value is returned
as the address of the accepted socket.