Update socket address handling for socket_connect events.
The compression of the address of non-IPV4 and non-IPV6 sockets
was updated to properly handle the length of the pathname
specified for an AF_UNIX socket.  For the remaining socket types
the length of the field being compressed was modified to use the
difference between the length of the socket address and the
header on the sockaddr structure.

Consistent with a new strategy the allocation of the digest
pointer handle was moved into the function that is creating the
digest value.