Remove special case for computation of a generic cell value.
Previously a randomly generated hash digest value was being used
for the computation of the CELL identity for a generically
modeled event.

Given that the Linux kernel maintainers do not like 'magic'
constants, convert the internal TMA to use the zero message digest
value for the computation.

The use of this value will eventually be deprecated as all of the
events are converted to explicit modeling.