Move pathname from event field to file field.
Previously the description of a file had contained a hexadecimal
presentation of the digest of the filename.  The length of the
pathname was also included.

In order to facilitate the implementation of load specific models
a file description now contains the pathname of the file define
by the 'path' key.  The pathname in the event{} description field
was dropped as it is redundant, and in may cases was not useful.