Eliminate SHA256 dependency on TPM PCR extensions.
The remaining instance of where the SHA256 digest size is hard
coded is in the loading of the PCR digest registers with the
security event state coefficient.

The new implementation compares the size of the digest register
with the hash function digest size that is in effect.  The
smaller of the two values is copied into the register.

This prevents extraneous material on the stack from being used to
extend the root model measurement.

In the case where the PCR digest register size is larger than the
security state coefficient, the register contents is zeroed
before the coefficient is copied into the register.  This
prevents inadvertent material from being extended into the
measurement.