Implement caching of aggregate values.
Rather than compute the hardware aggregate value each time it is
required the values are now cached with the digest function name
used as the index for retrieving the cached value.

This required the introduction of the following two functions:

trust.c:find_aggregate

trust.c:add_aggregate

The first function searches for a hardware aggregate value that
matches the digest function that is in effect for the modeling
domain.

The second function allocates and adds a hardware_aggregate
structure to the linked list that tracks the hardware aggregate
values that have been computed.

In testing the implementation an anomaly was noted in the fact
that the zero value hardware aggregate was always being used for
the root model.  This was being caused by the TPM not being
available when the first measurement in the root modeling domain
was registered.

The trust.c:trust_init() function was modified to be called at
the following initialization level:

device_initcall_sync

Which is the earliest that the TPM can be made available.

In turn this required the tsem.c:set_ready() be moved to the
following initialization level:

late_initcall

Which is after the TSEM TPM infrastructure has been initialized.

This in turn surfaced another issue that was causing the boot
process to hang.  This was secondary to the fact that the
following security event hooks were now being called before TSEM
was initialized to handle events:

inode_mknod

inode_setattr

inode_getattr

key_alloc

A check against the tsem_ready variable was added to each of
these hooks to cause them to be bypassed until TSEM is capable of
handling modeling events.