Propagate to task_id to newly allocated tasks.
The task_id structure that uniquely identifies the identity of
the executable code used for a process was in some cases being
pressed as all null bytes.  This is secondary to either a process
not invoking executing a new binary after fork or by a process
creating a thread.

Propagate the task identity when a new task structure is
allocated in order to have either of the above cases being
modeled as the same executable code.

This commit also removes a redundant return and conditional that
has no effect on how the function returns.