Use a single transform context per modeling domain.
In support of being able to handle security events that occur in
atomic context the TSEM code was modified to create and use
single hash context per modeling domain.  This eliminates the
need to allocate memory during the processing of an event itself.

The primary system include file, tsem.h, was modified to retain a
pointer that holds the name of the digest and a crypto_shash
structure that will be used for the modeling domain.  The
digestsize is now available from the hash specific tfm structure.

The inline tsem_digest() function now returns the pointer to the
transform structure for the hash function used in the modeling
domain, rather than the name of the hash function.  A pointer to
the name of the digest function is retained because the hashing
of digests and system aggregates uses the name to locate the
desired digest value.

As part of this code re-organization a number of functions no
longer pass the tfm structure to subordinate functions since it
is readily available through the tsem_digest() function in the
function itself.