Usem event description structure as list container.
Previously the tsem_trajectory structure that was exposed at the
global TSEM level (tsem.h) was used to create the list of
of events in the security description trajectory.  That structure
only contained a list_head structure and a pointer to the event
description structure.

This required a memory allocation in the add_trajectory_point()
function that was unneeded since the tsem_event structure can
easily carry the list_head structure and for the list itself.

The tsem_trajectory structure was completely dropped, as was its
memory allocation, by embedding the 'struct list_head' structure
in the tsem_event structure.  Appropriate updates to the
trajectory list traversals were made so they operated directory
on the tsem_trajectory pointer that is derived from the container
of the list_head.