Drop use of model state_point structure.
Similar to the issue that had been noted with the
tsem_trajectory structure the 'event_point' structure that was in
the model.c file was deemed to be unnecessary.  It tracks the
same information as the tsem_event_point structure that is used
as the backing storage for the security state coefficients that
have been generated.

While the tsem_event_point structure has an extra boolean value
in it to represent the known goodness of the point the amount of
data is negligible and there seems to be little utility to doing
an extra allocation for an 'event_point' structure.  This latter
structure is used to maintain the list of sorted events for the
purposes of computing the state value of the model.

Drop the use of the 'event_point' structure and convert the state
generation infrastructure to maintain and use the
tsem_event_point structure.