Implement support for internal modeling of atomic hooks.
This commit, rather extensive, lays the groundwork needed for
the modeling of security event hooks that are called with
non-sleeping locks.

The primary issue addressed in these changes is to implement
pre-allocated magazines for the following structures:

tsem_event

tsem_event_point

A similar strategy was used for each structure.  An array of ten
structures are allocated into static arrays in the event.c and
model.c files by the TSEM initialization routine.

When a 'locked' event is called the allocation for these
structures are made from the pre-allocated caches.  A request is
then dispatched to a work queue to refill these caches
asynchronously.

The following, globablly visible function, was added in order to
implement initialization of the model event points:

A bitmap is used to track which index in the array of
pre-allocated structures is allocated in order to designate the
slot to be refilled by the refill worker.

tsem_model_cache_init()

The following, globally visible function, is used by the TSEM
security event handlers to invoke the handling of locked events:

tsem_map_event_locked()

The tsem_allocate_event() function has an additional argument,
locked, that is used to signal to the event handler that the
event has to be handled in atomic context.  A 'locked' member was
added to the tsem_event structure in order for the rest of the
TSEM event handling infrastructure can determine how the event is
to be handled.

As a temporary measure, the following function was implemented in
order to selectively handle atomic context events:

tsem.c:model_generic_event_handle_locked()

The plan will be to use this function to handle the enrollment
of additional security events that are called from atomic
context.

The other major change is that the locks that protect the
security event trajectory and state points was changed from a
mutex to a spinlock in order to support handling events in
atomic context.

The critical sections protected by these slocks is very narrow,
with the exception of the routines in fs.c that have to traverse
the entire list but this is an infrequent operation only
permitted by process that hold CAP_TRUST.