Rename locks and convert forensics list lock.
The previous patch had left the naming of the locks with a _mutex
suffix which is now incorrect given that the locks have been
converted to spinlocks.  Rename the variables so the _mutex
suffix is changed to _lock.

The original atomic context patches had not converted the
forensics lock but that was an omission.  An invalid security
event for an event that is called in atomic context will generate
a lockdep warning.  Correct this by converting this lock to a
spinlock as well with the requisite name change.