Document memory barriers used in cache refill routines.
The checkpatch utility was raising an issue indicating that the
smp_mb__after_atomic() calls that are being used to force
visibility of modifications to the cache magazine index were not
documented.

Add documentation that indicates the calls are used to ensure
that the allocation and refills of the cache slots become
visible.