Enroll remaining locked security event hooks.
The security event hooks that were being bypassed since they were
being run in atomic context were enabled for internal modeling.

The exception is the tsem_unix_may_send() function, that is
causing the boot to fail so it is being bypassed until the reason
for the failure can be determined.  It appears that the failure
may be secondary to the TPM update being called in atomic
context.

The inclusion of the blocked security event hooks was implemented
by simply renaming the model_generic_event_handle_locked() to
model_generic_event_locked.  The tsem_unix_may_send() function
was modified to simply return a value of 0 since it is now the
only function called in atomic context that is bypassed.

Enabling the processing of the locked events was causing the
system to panic secondary to the tsem_key_permission() hook being
called before TSEM was initialized.  A tsem_ready check was added
to this hook in order to bypass processing the hook until the
system is ready to handle TSEM processing.