Separate event allocation from initialization.
In order to support changes that will be coming to support
asynchronous TPM updates the allocation of a tsem_event structure
was separated from its allocation and initialization.

A tsem_event structure is now allocated with the following
globally visible function:

tsem_event_allocate

This function takes a boolean argument to specify whether or not
the event structure is to be allocated in atomic or non-atomic
context.

The following function now replaces the tsem_event_alloc()
function:

tsem_event_init

Which takes the same function signature and is simply renamed
from the previous version.