Implement asynchronous TPM updates.
In order to complete the ability to handle security events that
are called with atomic context the extension of security event
coefficients into the TPM platform configuration register was
moved to an ordered asynchronous workqueue.

A 'struct work_struct' was added to the tsem_event structure,
along with an unsigned integer to hold the digest size of the
event.

The tsem_trust_add_event() function was modified to take a
pointer to a tsem_event structure rather than a pointer to the
coefficient that was to be extended into the register.

The extension of the TPM platform configuration register was
moved to a new function named tpm_update_worker().  The
tsem_trust_add_event() function schedules the event measurement
update onto an ordered workqueue that runs the
tpm_update_worker() function

The trust_init() function was modified to create the ordered
workqueue when TSEM is initialized.  An ordered workqueue is
used to ensure that the measurements are extended into the PCR in
the order that the event occurred.

The model.c:update_events_measurement() function was modified to
take a pointer to a tsem_event structure rather than a security
state coefficient.  This is to allow the structure pointer to be
passed to the tsem_trust_add_event() function.

The tsem_model_add_aggregate() function was modified to conduct
the extension of the platform aggregate into the TPM by
allocating a tsem_event structure pointer and populating it with
the model domain digest size and the platform trust aggregate
before submitting the structure to the tsem_trust_add_event()
function.