Simplify and correct handling of locked generic events.
Kernel segmentation faults were noted in the
model_generic_event_locked() function surrounding the handling of
the tsem_unix_may_send() event handler.

In the process of studying the problem it was noted that the
construction of the model_generic_event_locked() function was
sub-optimal and in ways incorrect.

The function was modified to return immediately on receipt of an
error pointer from the tsem_map_event_locked() function.

If the tsem_model_event() function returns an error code that
value is returned to the caller rather than invoking the
event_action() function.  A put on the tsem_event structure was
added to this function as this was missing and the event
structures would not be released when a subordinate modeling
domain was terminated.