Move work structure into tsem_TMA_context structure.
The work queue that is used to release the memory associated with
security namespace was using a structure named tsem_TMA_work to
encapsulate the modeling context structure.

This is unneeded if the work structure is embedded in the
tsem_TMA_context structure proper.  Move the 'struct work_struct'
structure into the tsem_TMA_context structure and adjust the
extraction of the context pointer appropriately.