Implement external modeling of locked security events.
This commit extends the ability to model security events that are
called in locked context to external modeling agents.

The primary structural change is to implement a cache of
pre-allocated structures used to form the linked list of events
that are to be read by an external modeling agent.  A strategy
similar to what was used for security event descriptions and
state structures was used.

The following externally visible function was implemented to
allow the TSEM initialization infrastructure to create a cache of
'struct export_event' structures that are used to export event
descriptions of security events called from atomic context.

The following static functions were added to the export.c file in
order to handle allocation and freeing of the export event
description structures:

free_export()

allocate_export()

Since event by event permissions checking is not available the
following JSON description is used to describe the export of an
asynchronous security event:

"export": {"type": "async_event"}

This allows a trust orchestrator to identify an event that was
called in atomic context and to exert appropriate actions on the
workload, such as shutting it down in the event of an out of
model event.

The remainder of the security event description is identical to
the descriptions for non-atomic events.

In order to support the atomic context constraint the mutex that
was used to protect the export list was converted to a spinlock.

The function signature for the tsem.c:model_generic_event()
function was modified to accept an additional boolean argument to
specify whether the function is called from standard or atomic
context.  This allowed the 'locked' variant of the function to be
removed and only a single function for handling generically
modeled events is retained.

The security event and export structure cache sizes was increased
to 64 as a result of testing the use of external modeling.  The
previous values were demonstrated to be insufficient.

It is important to note that the size of the event and export
caches need to track one another.  This is secondary to the fact
that for an externally modeled event the export structure holds a
reference to the event structure and thus reasonably equal
numbers of these structures are needed.