Update event and export cache sizes.
Testing of external modeling of security events indicated that
modeling was failing due to the event and export atomic
allocation magazines being depleted.

This commit expands the magazine size to a value of 96 that
consistently tests as successful.

Important to note is that running quixote-us with debugging
enabled increases the size of the cache that is needed to model
workloads without memory depletion errors.  A value of 64 is
sufficient without debugging but the level needs to be expanded
to 96 when debugging is enabled.

Also important to note is that the event and export description
structures need to be expanded in unison.  If they are
unbalanced, testing indicates that a workload will security fault
due to whatever cache description has the lower size.