Implement per model event point magazine.
Following in the pattern of the per security modeling domain this
commit implements support for a per model event point cache for
security events that are called in atomic context.

The creation of the per model cache is implemented in the
tsem_model_allocate() function and thus does not need an external
call point.

The tsem_model magazine_free() function needs to be visible
outside the model.c file since it has to be called in the event
of a failure to allocate the root model.

The allocation of the magazine for the root model is done in the
tsem_model_cache_init() function.  As a result the function
signature for this function needed to be updated to access a
pointer to the model description structure that is being
initialized.

The tsem_work struct was modified to use a union to hold the
pointers to the tsem_context and tsem_model pointers so that the
tsem_work structure is re-usable between the two magazine
implementations.  This will be extended to include the
tsem_export structure when per external TMA magazines are
implemented.