Implement per domain export caches.
The last issue to be addressed with respect to per modeling
structure magazines is to implement a magazine of structures for
the structures needed to support the export of security event
descriptions to external modeling agents.

The following, externally visible, functions were added to
support this:

tsem_export_magazine_allocate()

tsem_export_magazine_free()

The architecture implemented is very similar to the
infrastructure used for the event description structures and the
internal model points.

The magazine information for the structure caches is embedded in
the tsem_external structure that is allocated for an externally
modeled domain.