Fix hang on invocation of second modeling domain.
Testing indicated that if a second modeling domain was started
that domain would hang consuming CPU cycles and was
uninterruptible.

This was tracked back to the fact that when multiple digest
implementations were added the size variable that determines the
length of the memcmp comparison function was not being
initialized to the digest size.  This causing the memcmp to
always fail and the loop over the current task authentication
keys would continue endlessly.

Fix the problem by setting the size of the comparison to be the
size of the digest currently being used.