Implement trajectory occupancy counts.
One of the issues that will be needed to conduct machine learning
based models is a determination of the number of times a
particular security has occurred.  This commit surfaces that
value in the following pseudo-file:

/sys/kernel/security/tsem/trajectory_count

Each line in this file matches the corresponding line in the
/sys/kernel/security/trajectory pseudo-file.

A new cohort of functions was added to the fs.c file to support
the functionality of the pseudo-file.

The following functions in the model.c file:

have_point()

add_event_point()

Were modified to return the pointer to the security coefficient
point that was either found or added.  This allows the count
value to be independently incremented by the code that either
locates or adds a point.  This is needed since the code that
loads a security model needs to locate and add points without
incrementing their occupancy counts.

This allows attestation code to verify that a point, loaded by a
model, has been 'touched'.