Move points file to new location.
As part of the partitioning of trajectory points and counts into
valid and invalid lists this commit moves the summary of security
coefficients from the following pseudo-file:

/sys/kernel/security/tsem/points

To the following pseudo-file:

/sys/kernel/security/tsem/trajectory_points

In contrast to the previous 'points' file the latter pseudo-file
only outputs points that are valid.  The previous file had both
coefficients for valid and invalid events.