Implement per domain cache sizes and updated control plane.
This commit completes the process of having a separately sized
and independent set of caches for each modeling domain.  This has
the dual effect of allowing the cache sizes to be optimized for a
workload and prevents an adversarial workload from executing a
denial of service against another modeling domain.

Using positional parameters for writes to the control file,
particularly for namespace configuration was problematic and
would be criticized as problematic with respect to ABI changes.
In order to address this issue writes to the
/sys/kernel/security/tsem/control pseudo-file were modified to
use a key=value paradigm for the arguments needed to create a
modeling domain.

This opportunity was also used to convert the open coded
traversal of the array of control commands to the use of the
match_string() function supplied by the kernel library
implementation.  Properly support this required the enumeration
type that defines the type of control messages to be re-ordered
to start with an index of zero rather than 1.

The match_string() function was also used to determine the
argument type being passed to the config_namespace() function.
The index element located is used in a switch statement to evalue
the argument to the keyword.  Default values for the digest type,
namespace reference magazine size were implemented in order to
minimize the arguments that are needed to create a modeling
domain.

The following keywords are used to describe the arguements to the
namespace creation command:

nsref
digest
key
cache

Using to specify the argument for the namespace reference that
will be used, the digest function that is to be used for identity
compression, the key value for authenticating an external
trust orchestor to processes in the domain and the cache size for
atomic security events.

With this infrastructure in place it was straight forward to
convert the calls in the tsem_ns_create() function to use the
cache size, that is now passed as an argument to the function,
rather than the static value of the pre-processor define.

In order to facilitate clarification the enumeration type for
specifying that the DAC values are to be referenced against the
initial user namespace was changed from:

TSEM_NS_INIT

To:

TSEM_NS_INITIAL

Interestingly, testing with the new architecture suggested that
previous default value of 96 slots for the event and export
magazines was too large.  The implementation now uses a standard
allocation size of eight slots for all three magazine types.