Allow creation of namespaces with no arguments.
With the implementation of the key=value namespace arguments
default values were set for all of the arguments, with the
exception of the authentication key value for an external domain.

Short circuit the parsing of namespace arguments in the case of
an internal modeling domain with all default namespace
parameters.