Use mutex locks for control plane list traversals.
In order to support modeling of security hooks that were called
in atomic context the mutex locks that were used to replace the
various model lists were replaced with spinlocks.  Kernel decorum
suggests that spinlocks should be held for only a few C
instructions.

A strategy was implemented to migrate from a spinlock protection
to mutex protection by capturing the end of the list under a
spinlock for the appropriate list.  The spinlock is dropped and a
mutex lock taken that protects the end of the list.  The
remainder of the list processing occurs under the protection of
the mutex.

This new architecture is enabled by the fact that, for a given
model implementation, only additions are made to the list.
Additions can be made while the mutex protected list traversal
operation is conducted but any of those additions will not be
reflected in the transaction that is being conducted against the
control plane file.