Drop trajectory and forensics count.
The internal trusted agent model was maintaining a count of both
the number of trajectory and forensics points.

The trajectory value was not being used and the forensics value
was being used to set a limit on the maximum number of forensics
state points that could be accumulated.  This count was compared
to an arbitratry static value and had been thought to be useful
for prevent a denial of service attack back in earlier
implementations of TSEM.

The static limit would be criticized by the reviewers so will be
removed until when and if the issue of a denial of service attack
against the model would be considered relevant.