TSEM global include file documentation.
One of the drive by criticisms on the initial review of TSEM was
that the tsem.h include file had been dumped and the individual
structures should get introduced as they are used.

This is problematic secondary to the fact that the enumerations
and structures are included across most of the constituent
components files.

The strategy I am going to move forward with is to copiously
document the enumerations and structures in the include file.
The case can be made that the driver can be best understood after
having a solid grasp of the data structures that are manipulated
by the driver implementation files.