Special case root magazine size.
Testing with a full Ubuntu implementation indicated that the
default magazine of eight slots failed to support an error free
boot of the system to operational status.  This is presumably due
to the threaded nature of systemd causing the general of large
numbers of security events in atomic context.

Testing indicated that a magazine size of 96 entries was
sufficient to produce an error free boot.  This is in excess of
what is needed for security modeling domains so make the 96 entry
default value only valid for the root model.

Since it appears that the root magazine size may need to be
modified based on workload this commit introduces the following
kernel command-line option to configure the size of the root
domain magazine:

tsem_cache

Which takes an integer value that reflects the number of event
magazine slots that are to be created in the root model magazine.