Bypass processing of generic events in non-root modeling.
Testing of the changes to the root magazine size indicated that
when modeling of generic events was implemented the check for
non-root modeling mode was not incorporated in the
tsem.c:model_generic_event() function.

This function was updated to use a check similar to the code in
the model_event() function in order to bypass modeling of the
event in a TSEM mode 1 configuration.