Drop maximum forensics counts.
Previous versions of the code had maintained a count of the
maximum number of forensics events that would be admitted for
each modeling instance.

Reaction to this value was removed but the variable used to
maintain the count had not been removed from the modeling
structures.

Drop all references to this count value.