Do not propagate task key for root modeling namespace.
Previously the tsem_task_alloc() function was universally copying
the task_key from the parent task to the subordinate task.  The
propagation of this key only needs to occur for processes that
are running in a subordinate modeling domain.  The tasks in the
subordinate domain need to have a copy of the namespace key in
order to compute the authentication key that is verified against
the task_key in the trust orchestrator task.

Belts and suspender programming suggest that the task key should
not be propagated into any processes that are spawned from the
trust orchestrator, since the trust orchestrator control process is
the only entity that should be able to control processes in a
modeling domain.