Implement explicit modeling of task_kill.
In initial implementations of TSEM the task_kill hook was
explicitly modeled.  This transitioned to generic modeling when
the issue with calling security hooks in atomic context was
noted.

With the implementation of full support for handling security
hooks invoked in atomic context the task_kill hook was returned
to explicit modeling.

An issue was noted on shutdown with a kernel segmentation fault
being caused by what appears to be the inability to allocate a
tsem_event_point structure for modeling the final kill command.
This was addressed by strengthening the check for successful
completion of the modeling by checking that a valid
tsem_event_point was allocated, in addition to checking the
return code for the addition of the point to the security model.

With support for process authentication, the previous checks that
were implemented for preventing cross model signals was relaxed,
since the event description includes an indication of whether or
not the signal was a cross-model signal.