Properly handle reporting of security actions in atomic context.
Testing indicated that lockdep errors were being reported by the
tsem_export_action() function when untrusted security events
being called in atomic context were encountered.

This was secondary to the tsem_export_action() function not being
aware of the context in which it was being called.  Address this
by passing a 'locked' argument to the tsem_export_action()
action.  The kmem_cache_zalloc() function was replaced with the
locally defined allocate_export() function that is capable of
handling both standard security events and those called in atomic
context.

This change required the tsem.c:returned_trapped_task() and
tsem.c:returned_trapped_inode() functions be modified to accept
an argument describing the context that it is being invoked from.