Handle TSEM ready checks in a unified fashion.
The check for whether or not TSEM is ready to handle security
events should be called before any action is taken by the
security hook.

Some of the instances of the check were being called after the
test for whether or not the task is trusted.  Move these checks
to before the task trust status is checked.