Standardize magazine size declarations in TSEM header file.
The default magazine sizes for security events called in atomic
context were in the tsem.h file for subordinate internal and
externaly modeled domains.  The default magazine size for the
root modeling domain was defined in the tsem.c file.

Move the default root magazine size into the header file and
standardize on the naming.