Fix magazine cache size configuration and reporting.
Anomalies were noted in the configuration of the magazine sizes.
It became apparent that the tsem_model_allocate() function needed
to take a size specification for the size of the cache, it was
previously using a default size and that appears to be inadequare
and suffered from 'special treatment' of that cache size.

Add an argument to this function that allows the cache size to be
explicitly specified by the namespace creation code.

In a similar fashion a size argument was added to the
tsem_model_cache_init() function to specify the size of the
magazine to be allocated.  The TSEM initialization code now
passes the size of the root magazine based on its default value
and/or a modification by a kernel command-line argument.

The code that allocates the structurs with backing magazines was
re-structured to provide more informative reporting on an
allocation failure.

Allocation failures now report in the following format:

PROCNAME in NS_ID failed ALLOCATION_TYPE, cache size=NN

Where PROCNAME is the current->comm string for the process
requesting the allocation.  The NS_ID is the context identity of
the security modeling namespace that the request is being made
from, the ALLOCATION_TYPE is a description of the structure being
allocated and NN is the current cache size.

This format allows diagnosis of what process and modeling domain
is running into memory allocation problems and the size above
which the cache needs to be sized to.