Remove extraneous allocation of a security state point.
When the generation of the state value for a model was converted
from using an independent list implementation to a sorted array
the allocation of the event point for the state list had not been
deleted.  Remove that allocation.

Additional testing indicated that the root magazine size needed
to be raised from 48 to 96 entries.  This is a value sufficient
to allow a boot to normal operational status for a Debian
implementation.

An interesting note is the fact that it appears as if modeling
namespaces, both internal and external, on the Debian
implementation need far lower cache sizes.  This is a virtual
machine on newer hardware so this may be performance related.

This testing suggests that the virtual machine being tested on
should provide sufficient headroom for verifying the minimum
cache sizes that are needed.