Properly handle errors return from security point allocation.
The model.c:alloc_event_point() function returns an allocation
failure in the form of a error return code encoded in the pointer
that is returned.

Code paths that were calling this function were simply checking
whether or not a null pointer was being returned rather than
issuing the IS_ERR() function to check for an encoded error
value.  Fix the call points to use IS_ERR() and to extract the
error return value that is to be passed upward in the call stack.