Rationalize return characteristics from magazine allocators.
This comment standardizes the handling of the three structure
allocators that either allocate their structures from structure
specific caches or from the magazines that have been implemented
for security hooks that are called in atomic context.

All of the allocators now return a pointer to the allocated
structure on success or a NULL pointer if the allocation fails.

It is now up to the callers of the allocators to encode an
appropriate error return in the pointer if they are returning
pointers or to return an appropriate numeric error code.

The addition of either a trajectory or forensics point was
significantly simplified by taking advantage of the now
standardized behavior.