Place socket_accept address information in a union.
Completion of the documentation of the TSEM master header file
susggested that the implementation of the following structure:

struct tsem_socket_accept_args

Was inconsistent with the structuring of the
tsem_socket_connect_args with the former structure being wasteful
of space, given that all of the address information was
implemented as separate members of the structure.

The use of this structure was re-architected to include the
address information in a union contained within the structure.
The family information is then used to determine which member of
the union should be referenced.

The struct unix_sock member of the structure is only needed as
input to the event characterization code.  The path to the UNIX
domain socket is created in an on stack buffer in order to avoid
'clobbering' of the address in the union.

After the extraction of the path is complete the pathname is
copied into the path member of the union and thus overwrites the
pointer to the socket address information.