Update context status of security hooks.
Testing indicated that the ptrace_traceme was generating a
lockdep dependency error indicating that the modeling was being
done in an improper context.

This prompted a survey of the call points for all of the security
hooks to verify that the modeling context was correctly
configured in the TSEM security event handlers.  This survey
caused a number of the hooks, primarily task and IPC related, to
be changed from their standard process context to one that
modeled the event assuming atomic context.

The following two CPP defines were implemented in order to make
the code more self-explanatory:

LOCKED

NOLOCK

Where LOCKED is used to signal the caller that the security event
hook is running in atomic context, with the NOLOCK define
indicating that the event is running in standard process context.