Create independent directories for internal TMA's.
After some reflection on the comments that were received from the
first submission of TSEM the decision was made to attempt to
future proof the implementation.

Given the Linux policy of setting all interfaces in stone when
code enters the kernel there is a tension between surfacing
features and the need to implement the least common denominator
that will never be changed.

In order to allow additional modeling interfaces the access to
the model parameters for the in kernel TMA were moved from the
following pseudo-directory:

/sys/kernel/security/tsem

To the following, newly created, pseudo-directory:

/sys/kernel/security/tsem/InternalTMA/model0

The InternalTMA directory is designed to be a directory that
holds TMA implementations, just like the ExternalTMA
pseudo-directory is used to hold references to external trusted
modeling agent implementations.

Inside the InternalTMA directory the model0 directory is used to
hold the pseudo-files that provide access to the model
information.  This will allow additional trusted modeling agent
implementations to be surfaced by alternate pseudo-directories in
the InternalTMA pseudo-directory.