Document additional kernel parameters.
This commit updates the kernel parameters to reflect the kernel
parameters that were added after the tsem_mode command-line
parameter was introduced.

The following two new parameters were added:

tsem_cache
tsem_digest

The tsem_cache is used to specify the number of entries in the
structures that are pre-allocated to handle events in the root
security modeling namespace.

The tsem_digest parameter is used to specify the cryptographic
hash function that is to be used to create security event
coefficients in the root modeling namespace.