Enforce digest based size for external authentication key.
As part of the process of reviewing the TSEM documentation for
the V2 release it was noted that a statement is made that the
authentication key provided by the trust orchestrator for an
externaly modeled security modeling namespace must match, in
binary length, the size of the cryptographic digest function
being used for security state coefficient generation.

The only check being made was for a non-null key value.  Add a
verification check in tsem_ns_create() when the externally
modeled namespace is being created to verify that the key length
matches the digest size.