Use key=value arguments for trust control commands.
Paul Moore the Linux LSM maintainer has come out with an advisory
with respect to the requirements that must be met for having an
LSM upstreamed into the Linux kernel.

One of the requirements is an admonition that any control planes
must contemplate a 20+ year ABI stability requirement.  Given
that a lot of last minute consideration was given to the TSEM
control plane.

The most amount of flexibility with respect to command-line
arguments is if they are encoded as key=value pairs.  That
removes position dependencies on the arguments and allows control
plane requests to be extensible.

As a result all arguments to the 'control' file will be provided
as key=value pairs, even if there is a single argument to the
control keyword.

This commit starts this process by addressing the 'trusted' and
'untrusted' control verbs.  The command now take the following
form:

trusted pid=PID key=HEXID
untrusted pid=PID key=HEXID

Where PID is the base 10 ASCII representation of the process
identifier that will be acted on.  The HEXID is the ASCII base16
representation of the authentication key that was set for the
security modeling namespace when it was created.