Update MAINTAINERS file.
One of the review comments from the V2 release was the fact that
the entry for TSEM was not in alphabetical order in the file,
correct that fact.

In the process of doing this I noted that the location for the
ABI file was still named tsemfs rather than tsem.  Change the
MAINTAINERS file to reflect this correct path.