Correct documentation of kernel parameters.
An additional review comment for the second release of TSEM
surrounded the kernel parameters that were added to the following
file:

Documentation/admin-guide/kernel-parameters.txt

The three kernel parameters had not been order alphabetically so
the three entries were appropriately ordered according to
alphabetical ordering.

There was a desire for increased clarity surrounding what
tsem_cache meant.  I updated the documentation to very
specifically indicate that the size was the number of entries in
the arrays of pointers to structures that will be used to model
security events that occur in atomic context.

Finally there was a grammatical error for the tsem_digest
parameter that was corrected.