Add support for task disambiguation.
Work with the data science consultant to the Quixote project
suggested the need to enhance the output of security state
definitions in order to improve the ability to train a machine
learning model.

One of the issues was an ability to differentiate separate
instances of security events that were generated by disparate
invocations of the binary that generates a specific task
identity.

This issue was addressed by implemented by having a 64-bit bit
global counter that is incremented on each successful completion
of the tsem_bprm_creds_for_exec() function.  This value is
included in the event JSON structure as the ttd: member.  An
acronym that stands for task temporal disambiguator, since it
identifies separate instances of execution of the executable.