Add event timestamping.
An additional recommendation from the data science initiative was
to provide times for when each event occurred.

An additional member was added to the struct tsem_event structure
to hold the time in nanoseconds when a security event was
processed.  The event field of the trajectory output was modified
to include a ts: member which contains the timestamp value.

The timestamp counter used was a clock that counts continuous
time since the last boot that includes any time that was spent in
suspend.  This is to allow a model to capture a possible scenario
where logging may have been disrupted for a long period of time.