Drop notion of generically modeled events.
Previously, events that did not have characterizing parameters
were classified as having a type of 'generic_event'.  The CELL
tag for the event was 'generic_event' and had a type key that
contained the actual name of the event.

The data team that is working on developing machine learning
models found this problematic and wanted to have all events being
treated as first class events.

This commit drops the notion of a 'generic_event' and has all
events being named for the security hook that engenders the
event.  The CELL value is the name of the security event with an
empty value, {}, associated with it.

This change allows the event type enumeration to be dropped from
the tsem_event_parameters structures.  Instead a boolean
variable, no_params, was introduced into the tsem_event structure
to indicate that the event has no characterizing parameters.