Implement export only domains.
In order to better facilitate machine learning models an export
only mode was implemented, initially for non-root security
modeling namespaces.

While an export only model could be implemented for externally
modeled domains, having an explicit export only mode increases
efficiency since the exports can be done asynchronously from the
invoking process, ie. no delay in waiting for what would be
always be a positive affirmation of the security event.

The implementation was largely straight forward and consisted of
defining a new control type of TSEM_CONTROL_EXPORT.

In order to simplify the code flow the allocate_export() function
was added to the namespace.c file to create the export
infrastructure for the export only security modeling namespace.

A boolean flag variable named export_only was added to the
tsem_external structure that is set to indicate that the security
modeling namespace will only be exporting security event
descriptions and not waiting for a model response.