Properly handle pseudo-filesystem bypass.
An anomaly was noted in that inode based security events were not
being modeled.  This was traced to improper detection of
filesystems whose files were to be bypassed by the bypass_inode()
function.

The top of this function had a test to determine whether or not
the inode being passed to the function represented a 'regular'
file, if it did not a return value to bypass the hook was
returned.

This was proper behavior for the file_open and mmap_file
functions that were the original reason for this test.
Unfortunately the inode passed for the inode_* functions was
typically a parent directory or other directory designation.
This resulted in these hooks being universally bypassed.

This was corrected by removing the S_ISREG test to the file_open
and mmap_file functions.

The bypass_inode() function was renamed to bypass_filesystem in
order to more correctly represent the purpose of the function.

Testing of TSEM against a Debian boot resulted in the boot
process generating a segmentation fault.  This was caused by the
inode based security hooks being run before TSEM modeling was
ready.

This was addressed by placing checks for the readiness of
modeling in the following security event handlers:

tsem_inode_create()

tsem_inode_link()

tsem_inode_symlink()

tsem_inode_mkdir()