Implement parent event tracking.
One of the issues raised by the data science group was the need
to explicitly track the relationship between security events.

This was implemented by adding the following keys to the event:
description structure:

p_ttd

p_task_id

The p_ttd value will contain the parent ttd value for a process
while the p_task_id value will contain the task_id of the parent
process.

These are both set in the tsem_task_alloc() function when a new
process is created.

In testing this implementation an anomaly was noted where a
process, init in this example, had a task_id that was different
than its other invocations.  This was secondary to the task_id
being computed by the tsem_bprm_creds_for_exec() function which
is invoked by a process preparing to issue an exec system call.

After some experimentation the security_bprm_committing_creds()
security hook was found to be a better point in which to
implement this functionality.  This hook is called from a context
where the process name (comm) is consistent with the CELL
identity of the executable that gave rise to the process.

The tsem_event and tsem_task structures were updated to include
the following members:

p_instance

p_task_id

Which contain the parent ttd and task identity values for the
task and a stored security event description.