Feed parent task identity into security coefficient mapping.
With the infrastructure in place to support parent task
identities that value can be integrated into the mapping of a
security state coefficient.

The map.c:get_event_mapping() function was modified to feed the
parent task identity as the first data block into the security
modeling namespace digest function.

With this change the formula for a security state coefficient is
as follows:

Coeff = HF(HF(EVENT_ID) || P_TASK_ID || TASK_ID || HF(COE) || HF(CELL))

The TASK_ID itself is generated with the following formula:

Coeff = HF(HF(EVENT_ID) || P_TASK_ID || NULL_ID || HF(COE) || HF(CELL))

This significantly increases the sensitivity of the modeling
since it causes the security state coefficients to be dependent
not only on the characteristics of the event but also the
characteristics of the process chain leading to an event.

For example previously, the truncation of a logfile by the root
user in a shell session would not generate a security state
coefficient different from the modification of a file by a log
daemon.  With this change a shell session truncating the file
would generate an alternate security state coefficient.

It should be noted that this change does not impose a temporal
ordering dependency on the security state coefficients.  They may
occur in any order but the event must be fathered by a task with
a specific parent task identity value.