Bypass task identity creation if TSEM is not ready.
Testing against the 5.15 Debian kernel and userspace indicated
that a kernel null pointer de-reference error was occuring on
boot.  This was secondary to the tsem_committing_creds() function
attempting to generate a task identity before the TSEM
infrastructure was fully ready.

Correct this by placing the standard unlikely check at the
entrance to the function.