Convert inode_getattr to explicit modeling.
From this point forward an initiative is being following to
convert all of the generically modeled events to explicit
modeling.

This commit begins the process by converting the inode_getattr
hook to explicit modeling.  In order to better generalize future
additions the event.c file was modified to implement the
extraction of inode specific metadata into the get_inode_cell()
function.

The get_file_cell() function uses this function to acquire the
inode metadata.

The get_path_cell() function was modified to take a 'struct path'
argument rather than a 'struct file' pointer.

The current implementation encodes the information about the
inode whose attributes are being retrieved into a JSON file
directive.  Consideration is going to be given to using a
JSON inode structure to hold inode information.