Implement explicit modeling of inode_setattr security hook.
The commit converts the TSEM implemention of the inode_setattr
event handler from generic to explicit modeling.

The characteristics of the inode to be changed are specified by
the iattr structure that uses the valid member to indicate what
components of the structure have valid information to be changed.
In order to obtain consistent state values all members of iattr
structure that characterize the event are set to zero, except for
those indicated to be valid.

A new strategy is entertained in this commit for handling the
parameters for the event that will be expanded in use as new
evenets are fully modeled.

The tsem_inode_setattr_args structure uses an anonymous union to
hold two structures named in and out.  The in structure contains
pointers to the arguments supplied to the tsem_inode_setattr
handler while the out structure contains data elements from the
input structure that are maintained for the lifetime of the
security event description.

Since a dentry structure is used to specify the object whose
attributes are to be set, a new function was in the event.c file
that creates a path description given a dentry structure.