Implement full parameterization of inode_getxattr.
The inode_getxattr was converted from generic to full modeling, a
process that entailed a number of infrastructure enhancements
that will be used by subsequent full parameterization
initiatives.

The fill_inode() function was added to the event.c file in order
to populate the tsem_inode_cell structure that was added by this
commit.  The tsem_file structure has been used to carry inode
information but future commits will convert to the use of the
tsem_inode_cell structure to carry inode information.

The tsem_inode_getxattr_args structure was added to carry the
event parameters using the 'in' and 'out' structures that was
previously described.

The following helper functions were added to the map.c file in
anticipation of subsequent use.  These are as follows:

add_str()

add_inode()

These were used to simplify the addition of the security
parameters to the generation process for the security state
coefficient for the event.

One of the issues that this commit demonstrates is the need to
consider handling of paths that are generated with the
dentry_path() function.

In an environment that has a file in a filesystem mounted on the
root filesystem the pathname will be as follows:

/rootfs/tmp/dotest

In contrast, the /tmp/dotest file in the root filesystem will
generate the following path:

/tmp/dotest

The /rootfs dentry name represents the presence of the directory
which the filesystem has been mounted on.

If the d_absolute_path() function is used to retrieve the
pathname for the /tmp/dotest in a mounted filesystem, or bind
mount, the path will be the expected value of /tmp/dotest.

This underscores one of the issues with pathname based security,
the notion that different paths can represent the same file or
different files can be represented by the same path definition.

This will be addressed by further improvements that will include
reference information that specifies where the path is derived
from.