Convert sb_pivotroot to explicit modeling.
This commit is a straight forward conversion of the sb_pivotroot
LSM hook to generic modeling.

The tsem_sb_pivotroot_args structure was added to handle the
incoming and maintenance arguments to the security handler.

The JSON cell description is as follows:

"sb_pivotroot": {"old_path": "pathname", "new_path": "pathname"}

Where pathname is subsituted with the absolute path to the
filesystem root the paths are in.

The CELL mapping value is defined as follows:

CELL = HF(length || old_path || length || new_path)

Where length is the number of bytes in the pathname.