Change order for mapping of coefficient for inode_getxattr.
In order to maintain better logical consistency in the generation
of the CELL mapping a policy will be followed where the arguments
are added to the compression function in the order that they were
given as arguments to the security handler.

The previous mapping implementation for the inode_getxattr was
feeding the input to the compression function in the wrong order,
a fact that this commit addresses.