Implement support for root based path descriptions.
One of the issues with using a path specification in the creation
of a security event description is the fact that paths are not a
precise description of a file.  This is the underlying rationale
for label based mandatory access controls.

In order to address this the strategy used by the TOMOYO LSM will
be followed where the description of a path also contains a
directive as to the root of that path.

For a block based device the device number of the block device
specified in the superblock for the root of the device is used.
For mounts that are not device backed the name of the filesystem
mounted is used.

The following new JSON event descriptions for a path were
implemented:

{"dev": {"major": "NN", "minor": "NN"}, "path": "pathname"}

{"dev": {"fstype": "fs name", "path": "pathname"}

The tsem_path structure was added in order to support retaining
the descriptions of a path in a security event description.

A null value for the fstype pointer is used as a flag to indicate
that the dev member of the structure should be used to construct
the description of the root.

To migrate to this model the following functions were added to
the event.c file for converting LSM parameter descriptions to
TSEM path descriptions:

static int get_root(struct dentry *dentry, struct tsem_path *)

static int fill_path(const struct path *, struct tsem_path)

In the map.c file the following helper function was added to
encapsulate the output of a path description:

static void show_path(struct seq_file *, struct tsem_path *)

In the map.c file the following functions were added as helper
functions:

static int add_u32(struct shash_desc *, u32 value)

static int add_path(struct shash_desc *, struct tsem_path *)

Moving forward all security event parameter compression
algorithms will use the add_u32 function for adding a numeric
value to the security state coefficient.