Convert tsem_inode_getattr to use path infrastructure.
Previously the tsem_inode_getattr had used the tsem_file
structure to carry information about the inode being referenced
by the struct path argument supplied to the security event
handler.

This commit converts this event handler to use a tsem_inode_cell
definition to define the inode whose attributes are being
retrieved and the tsem_path infrastructure to detail the path.

The tsem_inode_getattr_args structure was introduced in order to
carry the struct path variable into the event characterization
code and to carry the tsem_inode_cell and tsem_path structures in
order to retain the event information.

The mapping code was significantly simplified by taking advantage
of the add_path() and add_inode() helper functions that were
previously introduced.