Convert inode_setattr to new path/inode infrastructure.
The tsem_path an tsem_inode_cell structures were added to the
tsem_node_setattr_args structure in order to replace tracking of
the target inode of the security event handler with the tsem_file
member of the event structure.

The following function was added to the event.c function in order
to fill a tsem_path structure using a direntry structure:

fill_path_dentry()

Since one of the members of the changed attribute structure is a
filesystem size that can be a 64-bit value the following function
was added to the map.c file:

add_u64()

Which, of course, adds a 64-bit integer to the generation of the
security state coefficient.

The creation of the security state coefficient for the event was
simplified by using the various helper functions that have been
introduced.