Convert inode_getxattr to path/direntry infrastructure.
The tsem_path structure was added to the tsem_inode_getxattr_args
structure in order to retain the path information for the inode
whose attribute is being interrogated.

The get_inode_getxattr() function was updated to use the
fill_path_dentry() function to populate the path information
structure for this event.

The trajectory output description structure was updated to use
the show_path() helper function to display the object path.

The computation of the security state coefficient was updated to
use the add_path() function to add the path description to the
security state coefficient stream.  The mapping function was also
updated to generate the coefficient input stream in the order
that the trajectory information is displayed.