Use a single helper function for display of a tsem_inode_cell.
The trajectory output for CELL's that included a reference to an
inode were using open-coded output of the contents of the inode.

This functionality was consolidated into a helper function in the
fs.c file named as follows:

show_inode()

The display handlers that were using open-coded implementations
were converted to use of the new helper function.