Move output of event type into show_event function.
Previously the output of the CELL name for the event description
was being done in each of the CELL output routines, after the
display of the event characteristics with the show_event()
function.

Placing the call to output the CELL description at the end of the
show_event() function eliminates the need for an extra function
call in each of the output routines.