Convert tsem_inode_listxattr to explicit modeling.
The conversion of the handler for the inode_listxattr was
straight forward and followed the pattern of the other 'xattr'
conversions.

These conversions suggest that the characteristics of the 'xattr'
event handlers are similar enought that they do not warrant
individual argument structurs.  The conversion of this handler
re-uses the tsem_inode_getxattr_args structure for its input and
retention characteristics.