Convert tsem_inode_setxattr to generic modeling.
This commit completes the migration of the security handlers for
extended attributes to explicit modeling.

As the immediately preceeding commits have done the
tsem_inode_getxattr args structure was enhanced to support the
additional members that are needed to support this security event
handler.

Unique to this conversion is the encoding of the value of the
extended attribute to a Base64 encoded string.  There appears to
be no requirement for what will be the contents of the extended
attribute to be a null-terminated string.  As a result the value
is Base64 encoded in anticipation of the value potentially having
binary values that are inconsistent with an ASCII string.

The extended attribute value is passed as its native binary value
to the model mapping route for the event that converts it to
Base64 encoding for storage on the trajectory history.