Consolidate extended attribute parameters.
Given that all of the security event handlers for extended
attribute operations enjoy a reasonably common set of arguments
this commit consolidates argument carrying into the following
structure:

struct tsem_inode_xattr_args

The event.c:free_cell() function was simplified in order to
consolidate the freeing of all the extended attribute information
into a single switch clause.

A routine checkpatch scan detected style errors in map.c that
were corrected

In addition the value member of the 'out' structure was
explicitly zeroed after the conversion of the attribute value
into Base64 format so that the free_cell() function does not act
on a previously freed pointer.