Drop notion of fstype for root descriptions.
The commit uses the ability of pseudo-filesystems to provide
path descriptions for files that are accessed on the filesystem
to provide the pathname for dentry or path specified inode values.

If the superblock does not provide a device description the
dentry->d_op->d_dname method is used to provide a path definition
for the dentry or inode that is being accessed.  If this method
is not defined the following pathname is used:

nodev:/

This commit also changes the key value for the path name in the
'path' structure from path to pathname.  This is designed to
avoid confusion between the specification of the path structure
from the actual pathname specified by the structure.