Consolidate file handling into common infrastructure.
Previously, due to historical issues in the codebase, security
events that took a 'struct file' pointer were using the tsem_file
structure to save the characteristics of the file.  The tsem_file
structure was outside the CELL union that contains the other
parameters of the security event.

The tsem_file structure was converted to the tsem_file_args
structure.  This latter structure was moved into the CELL union,
both independently and as a member of the tsem_mmap_file_args
structure.

The event.c file was updated to remove the add_file_name() and
get_inode_cell() functions.  The add_file_name() function is no
longer needed since the hash of the pathname is now done by the
map.c:add_path() function since the tsem_file_args structure
contains a tsem_path definition.

The functionality of get_inode_cell() has been previously
replaced by the fill_inode() function.

The mapping of the CELL coefficient was updated to separate the
handling of TSEM_MMAP_FILE from TSEM_FILE_OPEN and
TSEM_BPRM_COMMITTING_CREDS events.  This allowed the case of an
anonymous memory mapping to be handled correctly since an
implicit assumption cannot be made that a file definition is
available.

The output map.c:add_file() function was added to consolidate the
generation of the security coefficient for the description of a
file into a place that can be used by both the TSEM_MMAP_FILE and
TSEM_BPRM_COMMITTING_CREDS events.

The model.c:tsem_model_has_pseudonym() function was updated to
take the pathname of the file as an argument rather than the
tsem_file structure that it had previously taken.  This change
was also propagated downward into the
model.c:generate_pseudonym() function.

The encoding of a security event description into JSON was
modified so that the TSEM_FILE_OPEN and TSEM_MMAP_FILE handlers
used the helper functions that have been previously developed to
output the path and inode descriptions.

The fs.c:show_file_open() function was added in order to take the
display of the TSEM_FILE_OPEN handler out of the switch statement
and make it more consistent with the rest of the event handlers.