Properly free pathname memory.
The work previously committed that converted TSEM to use common
infrastructure for handling file descriptions was improperly
implemented since memory allocated for the pathname was not being
properly freed.

Update the switch structure in event.c:free_cell() to add clauses
for the TSEM_FILE_OPEN, TSEM_MMAP_FILE and
TSEM_BPRM_COMMITTING_CREDS events to free memory allocated for
the pathname.

The need for the pathname member of the tsem_event structure was
eliminated by the file handling consolidation changes since the
pathname, if one is even needed, is now part of security event
description infrastructure.

This change required the need to use a memset() function to zero
the tsem_mmap_file_args structure.  If the event was for an
anonymous memory mapping there was no guarantee that the pathname
member of the tsem_path structure would be valid.

This underscored the issue that there is probably no longer any
need for the tsem_event_parameters structure and its handling.

Since the CELL union in the tsem_event description structure has
an argument that defines the characteristics of each event, those
descriptions can be used in the event.c file, since each argument
structure is separate into an 'in' and 'out' component.