Rationalize order of task characteristic.
The members of the tsem_event structure that characterize the
task that is requesting handling of a security event were
re-arranged so that the pid value and the description of the task
were located next to one another.

This places the instance and parent instance members of the
structure immediately after the member that indicates whether or
not the event was called in atomic context and is arguably a more
rational grouping of the structure members with respect to
function.