Eliminate the use of tsem_event_parameters structure.
This commit represents an extensive set of changes that involve
the replacement of the tsem_event_parameters structure with the
use of the CELL union in the tsem_event structure to carry both
the input arguments and their retention values.

The tsem_event_allocate() function is now called from the context
of each security event handler function.  The allocate function
was also modified to take the enumeration type of the event in
addition to the locked status.

This allowed the elimination of the tsem_map_event_locked
function since the tsem_map_event() function no longer needs to
make a decision on how to allocate the event structure.

The call to the tsem_map_event() function was added to the
tsem_model_event() function in preparation for having the option
of multiple models which may modify the security event parameters
prior to their mapping.

This commit also explicitly zeros the argument structure for the
security event being handled for instances of where the argument
structure has 'in' and 'out' structures.  This is make sure that
in the case of allocated pointers that the pointer value will be
null if the memory allocation was not done as part of the
argument setup.