Fully parameterize the tsem_inode_mkdir event handler.
The parameterization of the TSEM handler for the inode_mkdir LSM
hook follows largely as a replicate of the TSEM handler for the
tsem_inode_create event handler.

This is because the two events, from a security perspective, are
functionally equivalent in that they result in the creation of an
inode in a directory heirarchy specified by an inode.  TSEM
converts this into a unique security event coefficient by virtue
of the fact that the handler type is encoded in the security
coefficient.

This commit also converts the mode argument into display as an
octal value, consistent with how previous mode values are being
encoded.