Fully parameterize the tsem_inode_symlink handler.
The tsem_inode_create_args structure was modified to take a
character pointer in the 'in' and 'out' structures to hold the
name of the file that is being linked to in order to support
parameterization of this handler.

A clause for this event was adding to the get_cell_mapping()
function to compute the parameterized security coefficient for
this handler.

The event.c:get_inode_symlink() function was added to translate
the input arguments into the retention arguments.