Fully parameterize the tsem_inode_mknod handler.
A dev_t member was added to the tsem_inode_create_args structure
in order to handle the parameters passed to the inode_mknod LSM
hook.

A clause was added to the switch statement that selects the
strategy for generating a security state coefficient to handle
the extension of the parameters for the inode_mknod event.

The fs.c:show_inode_mknod() function was added to create the
security event description for this event.

The event.c:get_inode_mknod() function was added to translate the
arguments to the security handler to the values that will be
retained for the lifetime of the security modeling namespace.

Of note in this function was the implementation of a parameter
leveling implementation for this security event.  If the creation of an
inode is on a tmpfs based filesystem the UUID value of the
filesystem superlbokc is set to a UUID with all bits set, ie:

ffffffffffffffffffffffffffffffff

This is needed in order to have constant security state
definitions generated since the tmpfs implementation appears to
choose a random UUID each time a filesystem is mounted.

This functionality will be ultimately moved to model specific
code as that implementation evolves.
approximation
approximation