Fully parameterize the tsem_inode_killpriv handler.
This commit provides a straight forward conversion of the handler
for the inode_killpriv handler using the generic inode arguments
parameter structure.

The event.c:get_inode_killpriv() function was added to convert
the the dentry that was the subject of the call into its TSEM
inode and path components.

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function to generate the security
coefficient for the event.

The fs.c:show_inode_killpriv() function was added to generate the
security event description.

It has been difficult to generate a test case where this handler
is called so the full parameterization needs additional testing.
Most notably for the correct composition of the JSON message.