Standardize clearing of retention arguments.
It became apparent during the work on parameterizing the inode
security handlers that only the 'out' member of an argument
structure needs to be totally cleared after function local copies
of the 'in' arguments have been extracted.

Standardize all of the memset calls to only clear the 'out'
member of an argument structure.