Generalize arguments for inode based security handlers.
The parameterization of the handlers for inode based events
demonstrated that most of the handlers could be implemented with
a single common argument structure.

The tsem_inode_create_args structure was renamed to
tsem_inode_args in order to generalize the name beyond the
inode_create LSM hook.

In the CELL union the member name was reduced from inode_create
to inode.