Reference correct argument structure for file free.
Fully parameterizing the file_ioctl handler caused a kernel fault
when the tsem_event structure was being freed.  This was tracked
down to the fact that the pathname free was referencing the
inode_attr member of the CELL union rather than the file member.

Correct the free to use the correct member of the union before
parameterizing the file_ioctl handler, so the latter will
reference the correct structure.