Fully parameterize the tsem_file_ioctl handler.
A cmd member was added to the tsem_file_args structure in order
to handle the parameterization of the TSEM handler for the
file_ioctl LSM call.

A separate clause was added to the switch statement in the
get_cell_mapping() function in order to generate the security
state coefficient for a file_ioctl event.

The generation of the security event description is implemented
by the fs.c:show_file_ioctl() function.

The 'arg' variable passed to the file_ioctl handler is being
disregarded since testing indicated that deterministic
trajectories could not be generated.  The documentation for the
handler indicates that the argument could be a userspace pointer
that should not be manipulated by the LSM.