Fully parameterize the tsem_file_lock handler.
The conversion of the handler for the file_lock LSM hook was
largely straight forward and was able to reuse the infrastructure
from the tsem_file_ioctl handler.

The most notable change was to add code in the
event.c:get_file_cell() function to bypass the collection of the
file digest if the inode does not represent a 'regular' file.

The tsem_file_lock() handler was returning an error if the lock
was executed against a directory secondary to a failure caused by
trying to obtain the checksum of a directory.