Fix random return value from tsem_task_kill handler.
The commit that re-architected each of the handlers in order to
simplify their construction resulted in the tsem_task_kill
handler returning random values given that they were using an
uninitialized retn variable.

The two affected points are designed to return a value of zero so
this situation was corrected by specifically returning an integer
zero value.