Fully parameterize tsem_ptrace_traceme handler.
The fs.c:show_task_ptraceme() function was added to generate the
description of an event involving this LSM hook handler.

After significant consideration the decision was made to extend
the security state coefficient stream with the task identity of
the parent process that was passed as an argument to the handler.

The reson for this dilemma is that the task identity of the
parent process is already integrated into the generation of the
security state coefficient.  There seems to be little downside to
feeding the identity back in a second time and it makes the
handling of this event consistent with other events that feed the
CELL mapping into the coefficient generation.