Fully parameterize the tsem_task_setnice handler.
An integer value was added to tsem_task_kill_args to hold an
integer value, that in the case of this handler is the 'nice'
value to set the process priority to.

After that the conversion to full parameterization was straight
forward with the addition of the fs.c:show_task_nice() function
to display the description of the event.

An additional clause was added to the switch statement in the
map.c:get_celling_map() to generate the security coefficient for
the event.