Fully parameterize the tsem_task_setioprio handler.
The handler for the task_setioprio was parameterized with only
the addition of the following function:

fs.c:show_task_value()

Which is designed to be a generic function for displaying a key
name with a specific numeric value.  The intent will be to
convert other task security handlers that display a numeric value
to use this generic function.

The switch clause for generating the security coefficient was
shared with the TSEM_TASK_SETNICE handler.