Fully parameterize tsem_task_prlimit handler.
The parameterization of this LSM hook implementation required the
addition of the following new argument structure:

tsem_task_prlimit_args

The struct tsem_COE structure was modified to include the
securebits member that holds the corresponding member of the
struct cred structure.

The event.c:get_prlimit() function was added that takes a pointer
to the tsem_task_prlimit_args structure and tranfers the
credential information from the struct cred structures to the
struct tsem_COE structures in the out structure.

This function in turn calls the event.c:fill_creds() function
that was introduced to carry out the transfer of a struct cred
structure to the retention structure.

The fs.c:show_creds() function was introduced to output the event
description for a tsem_COE structure.  A future commit will
convert the display of the COE credentials to the use of this
function.

The fs.c:show_task_prlimit() function was introducted to output
the CELL description for the security event.

The map.c:add_creds() function was implemented to encapsulate the
extension of a tsem_COE structure into the security coefficient
being generated.  A new clause was added to the switch statement
in the get_cell_mapping() function that calls this extension
function.