Fully parameterize the tsem_task_setrlimit handler.
In order to allow the common task argument structure to be used
for this handler a union was introduced into the
tsem_task_kill_args structure in order to allow a signed and
unsigned value to occupy the same member space in the structure.

Two unsigned 64-bit integer values were also added to the
structure to hold the current and maximum rlimit values being
set.

A new clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to extend the
arguments to the task_setrlimit call into the CELL mapping.

The fs.c:show_task_setrlimit() function was added to output the
CELL description for this event.