Finalize generation of CELL value for task_prlimit.
Security state invariance testing after the parameterization of
the tsem_task_setrlimit handler indicated that consistent
security state values were not being generated.

This was secondary to the common error of not finalizing the
digest generation for the tsem_task_prlimit handler.  Correct
this by calling crypto_shash_final on the digest.