Fully parameterize the tsem_task_prctl handler.
The following structure was added to hold the arguments for this
handler:

tsem_task_prctl_args

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function to extend the arguments to the
handler into the generation of the security state coefficient.

The fs.c:show_task_prctl() function was added to output the
description for the security event.