Fully parameterize the tsem_socket_listen handler.
In addition to parameterizing this handler this commit lays
generic infrastructure for handling socket based events.

The tsem_unix_socket_args structure was renamed to
tsem_socket_args since there are multiple security handlers that
deal with one or more socket arguments.  An integer value was
added to this structure to handle security event handlers that
include some type of integer value, in this case the backlog
value.

A new clause was added to the switch statement in the
map.c:get_cell_mapping() function to handle the generation of the
security coefficient for the event.

In addition the add_unix_socket() function was renamed to
add_socket() in order to have a generic mechanism for extending
the characteristics of a socket into a coefficient generation
stream.

The event.c:get_socket() function was introducted to have a
generalized mechanism for converting a socket argument into its
TSEM retention equivalent.  The get_unix_socket() function then
became a wrapper around two calls to the new function.

The fs.c:show_socket_listen() function was introduced to generate
the security state description for this event.