Fill correct socket address retention value for socket_listen.
Inspection of the code indicated that a mistake had been made in
the specification of the address of the tsem_socket_create_args
that will receive the description of the socket that is passed to
the tsem_socket_listen handler.

The incoming address is held in the structure pointed to by the
socka pointer but the description was being transferred to the
sockb description.

Correct this by setting the output pointer to the socka
description.