Fully parameterize the tsem_socket_sendmsg handler.
The parameterization of this handler was somewhat involved given
the need to properly handle the somewhat edge case of proper
handling of the sendmsg variant of the system calls that send
data over a socket connection.  Exercising this functionality
requires a test utility written to send a connection less message
over a socket.

In order to support this modifications were made to the
tsem_socket_args structure to add a void pointer to the 'in'
structure.

A boolean member was added to the 'out' structure and a union
that wraps a sockaddr_in and sockaddr_in6 structure.  If an
address pointer is found in the msghdr structure that is passed
to the LSM hook the IPV4 or IPV6 address structure is copied into
the output structures and the boolean flag variable, have_addr,
is set to true.

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function that computes the security
state coefficient for the event.  The have_addr member of the
output structure is checked and the target address and port
number of the message are extended into the coefficient if the
have_addr value is set to true.

The fs.c:show_socket_msg() function was added to generate the
security event description for the event.

The event.c:get_socket_msg() function was introduced to convert
the arguments to the handler into their retention variants.