Fully parameterize the tsem_socket_recvmsg handler.
The TSEM implementation of the socket_recvmsg LSM hook re-used
the infrastructure that was implemented for the socket_sendmsg
hook.

The major notable difference is when a pointer to the originating
address is retained.

In the socket_sendmsg implementation it was sufficient to check
whether or not the msghdr->msg_name pointer was non-null.
Testing found this to be non-sufficient since the
msghdr->msg_name pointer was non-null but did not seem to contain
a pointer to address information that was valid.

Testing also found that when the msghdr->msg_name was a non-NULL
pointer the msghdr->msg_namelen variable was zero.  This
suggested that the kernel socket implementation was indicating
that the msghdr->msg_name pointer was not valid since the kernel
was indicating that the length of the address information was
zero.

The retention of the argument information was triggered by a
sequential check of the pointer status followed by a check of the
length of the address information.  If both of these conditions
evaludate positively the address is retained and will be included
in the event mapping and description.