Fully parameterize the tsem_socket_getsockname handler.
The full parameterization of this event also included a
correctness fix that impacts the TSEM_SOCKET_LISTEN handler.

The event.c:get_socket_argument() function was introduced in
order to handle the copying of a single socket argument to its
retention value.  This function properly zero's the output
structure.

The fs.c:show_socket_info() function was introduced to output the
characteristics of a single socket pointer.  The
fs.c:show_socket_argument() function was introduced as a wrapper
for this function that also outputs the event description.

A new clause was added to the switch statement in the
map.c:get_cell_mapping() function to generate the security state
coefficient for the event.