Fully parameterize the tsem_inode_setsockopt handler.
The parameterization of this event required an addition of the
'optname' member to the tsem_socket_args struture in order to
hold the option name argument delivered by the LSM hook to the
handler.

The fs.c:show_socket_setsockopt function was implemented in order
to encapsulate the generation of the security state event.

A clause was added to the switch statement in the
map.c:get_cell_mapping() function to implement generation of the
security state coefficient for the event.