Fully parameteriz the tsem_socket_shutdown handler.
The parameterization of this handler consumed the infrastructure
implemented for the tsem_socket_listen handler with some generic
improvements.

The following function:

fs.c:tsem_socket_listen

Was renamed to the following:

fs.c:tsem_socket_shutdown

And enhanced to take an argument to the name of the key that will
be displayed.  This allows the function to be re-used between the
two functions.

The TSEM_SOCKET_SHUTDOWN enumeration type was added to the switch
clause that was generating the security state coefficient for the
TSEM_SOCKET_LISTEN.