Fully parameterize the tsem_sb_statfs handler.
The next most called security handler without a full parameter
implementation was the sb_statfs LSM hook.

Since additional superblock related LSM events will be
parameterized the following structure was introduced to
encapsulate the passed and retained arguments for this event
handler:

tsem_sb_args

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to handle the
generation of the security state coefficient for this event.

The filling of the retained TSEM equivalents of the arguments
was done the event.c:get_sb_statfs() function that is being
introduced by this commit.

The fs.c:show_sb_statfs() function was introduced in order to
handle the generation of the security event description.

The tsem_sb_pivotroot handler had introduced a security event
description structure specific to that event but I believe that
functionality can be assumed by the structure being introduced in
this commit as a generic encapsulation of arguments provided to
superblock security events.