Fully parameterize the tsem_key_perm handler.
The tsem_key_args structure was enhanced by adding a key_ref_t
member to the 'in' structure and a total of five new members to
the 'out' structure that contain members values from the key
structure referenced by the key_ref_t argument.

A separate clause was added to the map.c:get_cell_mapping()
function to generate the security state coefficient for the
event.

The event.c:get_key_perm() function was introduced to transfer
values from the arguments supplied to the LSM handler to their
retention forms in the 'out' structure.

The fs.c:show_key_permission() function was added to generate the
security state description for the event.