Fully parameterize the tsem_sb_mount handler.
The tsem_sb_args structure was modified to add an unsigned long
flags member to the structure at large and two char pointers and
a struct path pointer to the 'in' structure.

The 'out' structure was enhanced to include two additional
pointers to retain the data pointed to be the two input
pointers.  The 'out' structure already contains a tsem_path
structure that can be purposed to hold the path denoted by
the new struct path pointer that was added.

A new clause was added to the map.c:get_cell_mapping() function
to generate the security state coefficient for this event.

The fs.c:show_sb_mount() function was implemented to output the
security event description for this handler.

The event.c:get_sb_mount() function was implemented in order to
conduct the conveyance of the arguments to the LSM handler to
their retention values.  Also implemented was a caluse in the
free_cell() function in order to free the path description and
the two arguments passed to the handler.

This work is demonstrating the need to address an issue with
establishing security state determinancy for security workloads
that are implemented with runc.

In order to address a security vulnerability with the
proc/self/exe pseudo-symlink.  The runc binary executes
two mount system calls that both reference a temporary file
in the following directory:

/run/runc/NAME/tmpname

Where NAME is the name of the runc container.

The first mount appears to be of the /proc/self/exe onto the
'tmpname' location.

A second mount command is then executed that has no device name
but a target name of the temporary file/directory location.

This second mount is followed immediately by an unmount whose
target has not been defined.

Since the security state coefficient is dependent on the path
target for the mount sequential container runs results in
different security state coefficients.  Further work needs to
be done to determine how to detect this pattern and level
it to a common security state coefficient that becomes
invariant.