Fully parameterize the tsem_sb_umount handler.
The parameterization of this handler was able to use the
tsem_sb_args structure without additional member inclusions.

The event.c:get_sb_umount() function was introduced in order to
translate the input arguments to the security handler into their
retention equivalents.

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to generate the
security state coefficient for the event.

Similarily the fs.c:show_sb_umount() function was added to
generate the security state coefficient for the event.