Set proper context status for tsem_key_permission.
The full parameterization of the tsem_key_permission handler, due
to a cut-n-paste error had ended up with the improperly atomic
context status being supplied to the tsem_event_allocation()
function in the handler.  This resulted in the lock debug code
generating warning messages.

Properly set the status in the event allocation code to address
this.