Fully parameterize the tsem_bpf handler.
The following TSEM global structure was added to support the
input parameters and their retention values for the BPF class of
security handlers:

struct tsem_bpf_args

A new clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to generate the
security state coefficient for the event.

The fs.c:show_bpf() function was introduced in order to generate
the security description for the event.

Since the relevant input arguments are two integers there is no
need to have a transfer function in the event.c file for the
event.