Fully parameterize the tsem_bpf_prog handler.
The tsem_bpf_args structure was changed so that it included two
structures named bpf and prog.  These structures were enveloped
in a union so that they would both occupy the same memory space
in the parent structure.

This allows different names to be used for the integer arguments
for the tsem_bpf and tsem_bpf_prog handlers while consuming the
same amount of space.

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function to generate the security state
coefficient for this handler.

The fs.c:show_bpf_prog() function was introduced to generate the
description for the security event.