Fully parameterize the tsem_bpf_map handler.
This security event handler was parameterized following the same
patter that was used for the previous two BPF based handlers.

The 'map' structure was added to the union in the tsem_bpf_args
structure in order to encapsulate the arguments that will be part
of the security state coefficient for this handler.  This
structure contains the map type member of the bpf_map structure
that is used as an argument to the security handler.  The second
member of this structure is the fmode argument to the handler.

An argument can be made that the key and value sizes are worthy
of inclusion in the coefficient stream but additional experience
with this security handler is needed before that decision is
made.  As is the case with function arguments, additional
arguments can be added but once present removing them is a
possible ABI break.

A new clause was added to the map.c:get_cell_mapping() function
to generate the security state coefficient for the event.

The fs.c:show_bpf_map() function was introduced to display the
security event description for the event.