Fully parameterize the tsem_move_mount handler.
The tsem_move_mount and tsem_sb_pivotroot handlers are
conceptually similar with respect to the need to characterize a
security event.  The tsem_sb_pivotroot handler was architected to
use an independent structure for its arguments.

The parameterization of the tsem_move_mount handler was moved
forward under the notion of using the tsem_sb_args structure.
All that was needed was to add a second path pointer to the
'in' portion of the structure and and a second instance of a
tsem_path structure to the 'out' section.

A separate clause was added to the map.c:get_cell_mapping()
function to generate the security state coefficient for the
event.

The event.c:get_move_mount() function was implemented in order to
transition the two path arguments to their TSEM retention
equivalents.

The fs.c:show_move_mount() function was introduced in order to
generate the security event description for the event.