Convert tsem_sb_pivotroot to tsem_move_mount infrastructure.
From a security event characteristic the sb_pivotroot and
move_mount LSM hooks are conceptually similar in that two paths
are available on which to gate a security decision.

Previously the tsem_sb_pivotroot implementation was fully
parameterized using the tsem_sb_args structure for the input and
retention arguments.  It was a straight forward conversion to
convert the tsem_sb_pivotroot handler to use the same
infrastructure.

This allowed the generation of the security state coefficient to
be consolidated under a single clause in the
map.c:get_cell_mapping() function.

In a similar fashion the previous infrastructure for the
tsem_sb_pivotroot handler was dropped and replaced with the
event.c:get_move_mount() and the fs.c:show_move_mount()
functions.