Zero pointer address argument for task_prctl.
Testing with runc based cartridge workloads was indicating that
sequential runs were generating different security state
coefficients for the tsem_task_prctl handler.

This was secondary to the fact that for the following two command
options:

PR_SET_NAME
PR_GET_NAME

The first argument after the option is a pointer to the buffer
that will either hold the process name that is to be set or the
buffer that the name will be written to.  Since the pointer
values are not stable over subsequent executions of the command
different security state coefficients were being generated that
disturbed a constant security state.

Introduce the event.c:get_prctl() function in order to handle
'leveling' of these pointer argments to a NULL value.  It is
anticipated that additional action on these parameters may be
needed in the future, hence the use of a new function.

The documentation for the testm_task_prctl_args structure had not
been properly updated when the handler was parameterized.
Correct the documentation for the structure so that it is
relevant.