Fully parameterize the tsem_quotactl handler.
There are currently two LSM handlers for quota related security:

tsem_quotactl
tsem_quota_on

The following structure was introduced in order to hold the input
arguments and the retention data for these handlers:

tsem_quota_args

A new clause was added to the switch statement in the
map.c:get_cell_mapping() function that generates the security
state coefficient for an event.

The event.c:get_quotactl() function was added to convert the
arguments to the tsem_quotactl into their retention values.

The fs.c:show_quotactl() function was implemented to display the
characteristics of the security event.