Fully parameterize the tsem_quota_on handler.
The parameterization of this event was able to use the
tsem_quota_args structure that had been implemented for the
tsem_quotactl handler.

A separate clause was added to the switch statement in
map.c:get_cell_mapping() function to generate the security state
coefficient for the vent.

The event.c:get_quotaon() function was implemented to retain the
components of the dentry that was the argument to the handler.

The fs.c:show_quotaon() function was implement to output the
security description for the event.