Implement function to test for event bypass.
This commit introduces the following function:

tsem.c:bypass_event()

That checks the set of conditions that may be met that would
cause the processing of the security event to be bypassed.  This
function returns the boolean true value if there is no need to
process the event.

Currently the checks for event bypass are in the following
functions:

dispatch_event()
dispatch_generic_event()

The issue with the checks there is that time has been spent on
the event allocation and setup that are completely unneeded.

A set of the security events that provide security verification
for network tunnel events do not have any arguments and will thus
be 'generically' modeled events that have no parameters.

Add the bypass event checks to these functions since their
parameterization will simply involve the allocation of an event
structure with the 'no_params' flag set in the event description.