Fully parameterize the tun handlers.
There are four LSM handlers that are involved in security
decisions regarding network 'TUN' devices that are as follows:

tsem_tun_dev_create
tsem_tun_dev_attach_queue
tsem_tun_dev_attach
tsem_tun_dev_open

Only one of these, tsem_tun_dev_attach, takes a descriptive
argument for the event.

The three LSM event handlers that do not have arguments were
converted from using the dispatch_generic_event() function to its
open coded equivalent.  This is to assist in reaching the goal
where all of the TSEM LSM event handlers all use a common
architecture for the event setup.

This approachalso takes advantage of the recently introduced
bypass_event() function so that no time is spent on an event if
the security modeling configuration does not require processing
of the event.

The tsem_tun_dev_attach handler has only as an argument a single
struct sock pointer.  As a result it is conceptually similar to
the tsem_socket_getsockname and tsem_socket_getpeername handlers.
In recognition of this the tsem_socket_args structure was used to
process and retain the argument to the handler.

The clause in the map.c:get_cell_mapping() function that was
handling the tsem_socket_getsockname event was re-used to
generate the security state coefficient for the event.

In a similar fashion fs.c:show_socket_argument() was used to
generate the security event description.