Fully parameterize the tsem_ipc_permission handler.
This commit initiates the process of parameterizing the security
handlers that make decision on IPC/shared memory operations.

The following structure was added to the global TSEM structures
to establish ownership for an IPC resource:

struct tsem_ipc

This structure encloses a byte array that is loaded with the task
identity of the process that is creating the resource.

The following structure was created to encapsulate arguments to
the shared memory handlers:

struct tsem_ipc_args

This structure follows the 'in' and 'out' architecture that has
been previously established as the design pattern for LSM
argument handling.

All of the IPC/shared memory handlers have as an argument a
pointer to a kern_ipc_perm structure.  In order to provide
argument retention support the following structure was
implemented:

struct tsem_ipc_perm

The following inline accessor function was also added to the
tsem.h file in order to return a pointer to the TSEM IPC
ownership structure:

static inline struct tsem_ipc *tsem_ipc(struct kern_ipc_perm *)

TSEM event handlers were added for the following events:

shm_alloc_security
sem_alloc_security
msg_queue_alloc_security

In order to direct these events to the creation of the TSEM
identity for these resources.

These functions are directed to the following event handler:

tsem_ipc_alloc

Which conducts the copy of the task identity of the caller into
the resource structure that is being created.

A reservation was also added to the tsem_blob_sizes structure for
the .lbs_ipc member that sets the size of the tsem_ipc structure.

The following two functions were added to the event.c file:

get_ipc_permission()
get_ipc_cred()

The first function handlers the transfer of arguments for the
tsem_ipc_permission handler.

The get_ipc_cred() function is a generic helper function that
transfers values from a kern_ipc_perm structure into the
tsem_ipc_perm structure.  This latter function will be re-used by
additional IPC handler implementations that take a kern_ipc_perm
argument.

In a similar fashion the map.c:add_ipc_cred() function was added
to extend the security state coefficient with the members of the
tsem_ipc_perm structure.

For the display of the security state description the following
generic helper function was implemented:

fs.c:show_ipc_cred()

Which outputs the IPC permission credentials, including the owner
of the resource.

The fs.c:show_ipc_permission() function was added to drive the
creation of the security state description for the
tsem_ipc_permission handler.  This latter function takes
advantage of the show_ipc_cred() helper function.