Parameterize multiple IPC/shared memory handlers.
A number of the security handlers for events involving shared
memory primitives have a common call architecture with a
pointer to a kern_ipc_perm structure and an integer flag or
control variable.

This commit fully parameterizes the following handlers:

tsem_shm_associate
tsem_shm_shmctl
tsem_shm_shmat
tsem_sem_associate
tsem_sem_semctl
tsem_msg_queue_associate
tsem_msg_queue_msgctl

The event.c:get_ipc_permission() function was already in place
that handled the translation of the kern_ipc_perm members to
their TSEM retention structure.  This function was re-used for
all of the handlers that were addressed with this commit.

In order to provide a generalized interface for outputting a
security event description for these handlers the following
function was introduced:

fs.c:show_ipc_shm_value()

This function was architected to take a character pointer that
provides the name of the integer argument to each of the
handlers.

Separate clauses were added to the switch statement in the
tsem_fs_show_trajectory() function that calls the
show_ipc_shm_value() function with the correct argument name.

The same clause in the switch statement in the
map.c:get_cell_mapping() function was used to generate the
security state coefficient for all of these handlers given that
a common architecture can be used for the generation of the
coefficient.