Fully parameterize the tsem_msg_queue_msgrcv handler.
A member named type was added to the tsem_ipc_args structure in
order to retain the type argument to this handler.  In addition a
pointer to a task_struct structure was added to the 'in'
structure to hold the target argument to the handler.  A u8
buffer was added to the 'out' structure to retain the task
identity associated with the target argument to the handler.

A clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to compute the
security state coefficient for this event.

The event.c:get_msg_queue_msgrcv() function was added to convert
the arguments to the security handler to their associated
retention values in the 'out' structure.

The fs.c:show_msg_queue_msgrcv() function was added to output the
security event description associated with the call to the
handler.

This commits marks the end of the conversion of all the TSEM
security event handlers to fully modeled implementations.  This
allowed the tsem.c:dispatch_generic_event() function to be
retired from use.