Convert tsem_socket_create to common argument infrastructure.
Since the socket security handlers had grown in two separate
phases there is a need to get the argument handling and naming
standardized.

It was a straight forward exercise to convert the handler for the
security_socket_create LSM hook to use the tsem_socket_args
structure rather than the call specific tsem_socket_create_args
structure.  The socket parameters passed to the call were copied
into the appropriate files of the socka structure in the 'out'
portion of the tsem_socket_args structure.

The clause in the switch statement in the
map.c:get_cell_mapping() function was converted to use the
add_u32() helper functions to extend the arguments into the
security state coefficient that is being generated.

The fs.c:show_socket_create() function was modified to use the
common tsem_socket_args structure for the generation of the
security event description.