Include kernel socket status in socket information output.
In pursuit of further standardizing the socket handlers the
fs.c:show_socket_info() function was modified to include the kern
member of the tsem_socket_create_args structure.

The event.c:get_socket() function was updated to set the value of
that kern member to the value specified by the sk_kern_sock
member of the struct sock structure that is passed as an argument
to the socket security handlers.