Convert socket_bind and socket_connect to common event arguments.
The tsem_socket_bind and tsem_socket_connect security handlers
use common infrastructure for modeling and mapping.  A review of
these handlers indicates that it is a straight forward
modification to convert them from using their own argument
structures to the use of the tsem_socket_args structure.

A path member and a mapping member were added to the union in the
'out' structure of the tsem_socket_args structure.  These will
hold the UNIX socket address and the mapping of socket addresses
that are not explicitly treated.

For the purposes of these two handlers the value member of the
tsem_socket_args structure was purposed to hold the length of the
socket address.  The addr member of the 'in' structure was
purposed to hold the pointer to the sockaddr structure that is
an argument to these two security handlers.

The pointer to the sock structure that defines the socket that is
the subject of these two handlers are placed in the socka member
of the 'in' structure so that it can be handled by the common
socket mapping and modeling infrastructure.

Due to line length considerations the generation of the security
state coefficient for these events was moved from a clause in the
switch statement in the map.c:get_cell_mapping() into the
add_socket_connect_bind() function in the map.c file.  This is
also consistent with how the event and display processing code is
organized in the event.c and fs.c compilation units.

The generation of the security event description in the
fs.c:show_socket() function was modified to extract the socket
information from the tsem_socket_args structure.  The most
notable change is that the show_socket_info() function is called
to output the socket characteristics.  This causes these two
handlers to generate standardized output with the other socket
handlers.

The event.c:get_socket_cell() function was re-organized so that
it takes a pointer to a tsem_socket_args structure rather than
a pointer to the previous handler specific argument structure.

The event.c:get_socket() function was renamed to
get_socket_address to reflect that the purpose of the function
was to return a mapping of a socket address.  A pointer to a
sockaddr structure was added to this function to supply the
address whose mapping is to be computed.