Implement explicit handling for dentry arguments.
The inode security handlers frequently take a point to a dentry
structure as an argument.  The TSEM relevant components of a
dentry are the backing inode and the path to the backing inode.

Previously, a dentry argument was represent by a tsem_path and
tsem_inode_cell structure.  This commit implements the
tsem_dentry structure to represent a dentry value with a common
TSEM structure.

In addition to containing tsem_inode_cell and tsem_path
structures the tsem_dentry structure contains a boolean value to
specify whether or not the tsem_inode_cell structure has been
populated.  This is to address the issue that in a number of the
handlers the dentry structure is only relevant for its path
description.

The event.c:fill_dentry() function was implemented to encapsulate
the transfer of the relevant contents of a dentry structure to
its TSEM equivalent.  This function determines if there is a
backing inode for the dentry and sets the flag variable and
populates the structure if a backing inode is available.

The event.c:free_cell() function was updated to reflect the fact
that the path structure components to be freed are now embedded
in a tsem_dentry structure.

The map.c:add_entry() function was implemented in order to
encapsulate the generation of the security state coefficient for
an event that accepts a direntry as an argument.

In a similar fashion the fs.c:show_dentry() function was
implemented to encapsulate the output of the dentry component of
a security state description.

Both the map.c:add_entry() and fs.c:show_dentry() functions check
the status of the have_inode flag and only incorporate or display
the inode if it is available.

This change improves the consistency of the security event
descriptions with a goal of having the key values in a cell
description represent the actual argument names to the security
handlers.