Convert extended attribute handlers to use dentry structures.
The tsem_path and tsem_inode_cell structures were dropped from
the out structure of the tsem_inode_xattr_args structure and
replaced with a tsem_dentry structure.

The following functions in event.c were updated to migrate the
dentry argument to the tsem_dentry structure for retention:

get_inode_setxattr
get_inode_getxattr
get_inode_listxattr

The following functions in fs.c were updated to use the
show_dentry() function to output the description of the dentry
that the security handlers are operating on:

show_inode_setxattr
show_inode_getxattr
show_inode_listxattr

A more significant change was made to the map.c in order to
consolidate the generation of the security state coefficient for
all four security handlers into a common function.  The following
function was added:

add_xattr()

To feed the security state coefficient input stream into the
compression function for the namespace.

This function tests the security event type in struct tsem_event
structure that is passed to it and returns after the parameters
relevant for each event have been fed into the compression
function.

The disparate clauses in the switch statement in the
get_cell_mapping() function were consolidated into a single
clause that calls the add_xattr() function.