Conver tsem_inode_rename to common inode argument.
A new_dir member was added to the 'in' structure of the
tsem_inode_args structure and new_dir member was added to the
'out' structure.  These new members are used to represent and
retain the new_dir argument to the handler.

The tsem_inode_rename_args structure was dropped from the tsem.h
file since it is no longer used.

The event.c:get_inode_rename() function was updated to retain the
arguments for this handler into the 'out' portion of the
structure using the fill_path_dentry() function.

The event.c:free_cell() function was updated to release the path
arguments from the dentry structures that they were propagated
into.

The generation of the security state coefficient for this event
was transferred from the clause in the switch statement in the
map.c:get_cell_mapping() function into the add_inode_rename()
function.  This function uses the add_dentry() function to
propagate the dentry components into the coefficient generation
stream.

The fs.c:show_inode_rename() function was updated to use the
show_dentry() function to output the dentry encoded arguments.