Change argument order on tsem_fs_show_key function.
A pattern has been established where functions that generate
portions of the security event description have the desired key
name first in the argument list, followed by the character
termination for the generation portion of the output.

The fs.c:tsem_fs_show_key() function had the opposite ordering.

Convert the tsem_fs_show_key() function to have the preferred
ordering for the key name and the termination sequence.