Update tsem_sb_umount handler to use dentry argument.
The superblock revelant security handles have not been upgraded
to use the tsem dentry infrastructure that has been previously
introduced.  Add a tsem_dentry structure to the 'out' structure
of the tsem_sb_args structure in order to transition relevant
handlers to use the dentry infrastructure.

The tsem_sb_umount handler was updated to use the dentry
infrastructure.

The event.c:get_sb_umount() function was updated, and simplified,
to transition the mount point description to a tsem_dentry
retention structure.  The free_cell() function in this
compilation unit was updated to free the path that was in the
dentry structure.

The map.c:add_sb_umount() function was introduced to encapsulate
the generation of the security state coefficient for this event.
This function uses the add_dentry function to add the dentry
components to the security state stream rather than the previous
code that used separate functions for the path and inode.

The fs.c:show_sb_umount() function was updated to use the
show_dentry() function to output the description of the dentry.
Rather than the name dentry for the field the field was named mnt
in order to match the argument to the handler.