Update tsem_sb_remount to use dentry infrastructure.
The event.c:get_sb_remount() function was updated to propagate
the dentry from the superblock into the tsem_dentry structure in
the 'out' structure of the tsem_sb_args structure.

The free_cell() function was updated in tandem to free the path
definition from the path component of the dentry structure.

The map.c:add_sb_remount() was introduced in order to encapsulate
the generation of the security state coefficient for the event.

The fs.c:show_sb_remount() was updated to encapsulate the
security descriptions of the superbock in a JSON structure named
sb. This is to continue the theme of having arguments to the
security handlers named for the argument.