Convert tsem_quota_on to use dentry infrastructure.
A dentry member was added to the 'out' structure of the
tsem_quota_args structure in order to allow the dentry argument
to be retain in a tsem_dentry structure.

The event.c:get_quotaon() function was modified to use the
fill_dentry() function in order to transfer paraments from the
dentry argument to the handler into the retention structure.  The
free_cell() function was updated to free the storage from the
path member of the tsem_dentry structure.

The clause in the switch statement in the
map.c:get_cell_mapping() function was updated to simply call the
add_dentry() function in order to generate the security state
coefficient for the event.

The fs.c:show_quotaon() function was updated to use the
show_dentry() function to output the security event description.