Convert tsem_quotactl to use dentry and superblock conventions.
The event.c:get_quotactl() function was updated to use the
fill_dentry() function to transfer the input dentry structure to
its retention value.  The free_cell() function in this
compilation unit was updated to free the memory for the path from
the path member of the dentry structure.

The map.c:add_quotactl() function was added to encapsulate the
generation of the security state coefficient for the event.

The fs.c:show_quotactl() function was updated output the
components of the superlblock in a JSON structure labeled with a
key name of sb.

This change allows the tsem_inode and tsem_path members to be
retired from the 'out' structure of the tsem_quota_args structure
since their role has been subsumed by the tsem_dentry structure.