Update handling of tsem_task_setrlimit handler.
One of the arguments to this security event handler is a pointer
to an rlimit structure.  Generate a structure output to enclose
the members of the structure are display with a key value of
new_rlim.

Move the code that generates the security state coefficient for
this event into the map.c:add_task_setrlimit() function.  In
doing so a regression in the coefficient generate was addressed.

Previously the TASK_ID of the task was being added into the hash
function using the size of the target member of the 'out'
structure in the tsem_task_kill_args structure.  This was causing
a stream of bytes equal to HASH_MAX_DIGESTSIZE being fed into the
compression function rather than the size of the TASK_ID which
will be the digest size of the hash function that has is being
used for the security namespace.