Update tsem_inode_setattr.
The second argument to this security handler is a pointer to an
iattr structure.  Consistent with the move to have the security
state descriptions mimic the arguments to the security handler
the fs.c:show_inode_setattr() function was modified to have the
members of the iattr structure enclosed in a JSON structure with
a key name of iattr.

The map.c:add_inode_setattr() function was introduce to
encapsulate the generation of the security state coefficient for
this event.