Update handling of tsem_settime handler.
In order to establish consistency with the ongoing work to
rationalize the security state descriptions the following members
were added to the tsem_time_args structure:

have_ts
have_tz

These boolean variables are set to indicate whether or not the
structure pointers to the ts and tz pointers have been set.

The map.c:add_settime() function was introduced to encapsulate
the creation of the security state coefficient for the event.
This function checks the status of the indicator variables and
only extends the coefficient generation stream with the contents
of the corresponding structure members if there is an indication
that the arguments were supplied to the handler.

The fs.c:show_settime() function was updated to display the
contents of the ts and tz structures as JSON encoded structurs
with the key names of ts and tz respectively.

As with the mapping function, this function checks the status of
the boolean variables to determine whether or not the
corresponding structures should be displayed.