Drop extraneous argument from tsem_mmap_file.
The reqprot argument had, at one time, been a relevant argument
to this handler.  A review of the code indicated that the
security_mmap_file function is being called with only three
arguments.

Interestingly the macro that defines the handler is still
requiring four arguments.

Rename the final argument to tsem_mmap_file to 'extra' to
indicate that it is unused, but needed, to get proper
compilation.

The reqprot member of the tsem_mmap_file_args structure was
dropped in order to support this change.

The fs.c:show_mmap() function was modified to drop the output of
the reqprot argument.  In a similar fashion the extension of the
reqprot argument was dropped from the map.c:add_mmap_file()
function.

Of particular note is that the freeing of memory associated with
this event was moved into its own clause in the switch statement
in the event.c:free_cell() function.  Previously the free of
the path was being called on the wrong event parameter.