Convert open coded generation of security state coefficient.
Code review during the implementation of the
tsem_ptrace_access_check handler indicated that the security
state coefficient for the tsem_ptrace_traceme handler was being
done by an open-coded hash call in a clause in the switch
statement in the map.c:get_cell_mapping_function.

Previously the map.c:add_task() function had been implemented in
order to encapsulate the extension of a task identity.  A call to
this function was implemented in the clause in the switch
statement that was handling the TSEM_PTRACE_TRACEM event.