Implement tsem_capable handler.
The security handlers that handle capability checks were deemed
important to justify their implementation before the next patch
series is released.  For example, the TSEM implementation for
security_capable guarantees that the check for a capability is
tied to a specific body of executable code.

The following structure was implemented to handle arguments for
this and subsequent security capability handlers:

tsem_capability_args

With an instance of the structure placed in the CELL union of the
evnet description structure.

The tsem_capable() function was implemented in tsem.c, along with
a macro call to initialize the security handler.

The name of the security handler was added to the tsem_name[]
array along with an additional enumeration element for the
tsem_event_type enumeration.

The map.c:add_capable() function was introduced to generate the
security state coefficient for the event from an additional
clause in the switch statement in the get_cell_mapping()
function.

In a similar fashion the fs.c:show_capable() function was
implemented to generate the security state description for the
event.  The function is called from a new clause in the switch
statement in the fs.c:tsem_fs_show_trajectory() function.