Implement tsem_capset.
This event handler was able to use the existing infrastructure
for the capability based security event handlers.

The event name was added to the tsem_names array and a new
enumeration value was created for the tsem_event_type.

The tsem.c:tsem_capset() function was introduced to implement the
handler.  A macro call to register the handler was also added.

The map.c:add_capset() function was implemented to generate the
security state coefficient for the event.  A clause was added to
the switch statement in the map.c:get_cell_mapping() function to
invoke the function.

The fs.c:show_capset() function was implemented to generate the
security event description for the event.