Integrate untrusted task handling into event dispatch.
Previously the detection of an untrusted event and its logging or
export to an external TMA was done by code that was implemented
in each security event handler.  These separate instances were
dropped and replaced with a call to the tsem.c:untrusted_task()
function in the tsem.c:dispath_event() function.

For the current time the tsem.c:trapped_inode() function was
dropped since this change resulted in only a generic log message
for models that were being serviced by the in-kernel TMA.

The full event description will be exported to an external trust
orchestrator but additional infrastructure will be needed to
log the full event description for internally modeled namespaces.

The tsem.c:event_action() function was removed since there is now
only a single call site and the simplification of the untrusted
task handling now means that only a single conditional is needed
in the tsem.c:untrusted_task() function to handle the decision on
what type of error code is to be returned.