Support invariant owner labeling of temporary files.
This commit is a major structural update to TSEM that lays the
infrastructure for handling files that are created in a security
modeling namespace in a manner that allows the functional value
to remain invariant between runs of a workload.

The most fundamental change is to add the following four members
to the tsem_inode_cell structure:

bool created
u64 creator
u64 instance
u8 owner[HASH_MAX_DIGESTSIZE]

The created member is a boolean used to indicate that the inode
has been created, ie. the product of the tsem_inode_create()
handler.

The creator member is the numeric identifier of the security
modeling namespace in which the tsem_inode_create() function was
called to create the inode.

The instance member tracks, on the basis of a TASK_ID, the number
of files that the task identity has created within a given
directory.  The owner member is the TASK_ID of the process that
created the inode.

The tsem_inode_owner structure was introduced in order to track
inodes that are created.  This structure is designed to support
linked lists of inodes that are created in a given directory.  It
reproduces the creator and ownership information that was
detailed for the tsem_inode_cell structure.

The inode information is also replicated in the tsem_path
structure so that instance information for the inode that is
backing a path can be displayed in either a trajectory or export
description.

The inode ownership information was also added to the tsem_inode
structure.  In addition the following three linked list
implementations were added:

create_list
instance_list
digest_list

Each list has its own mutex lock, ie. create_mutex, instance_mutex and
digest_mutex that protect the respective lists.

The create_list contains a list of tsem_inode_owner structures
that track inodes that are created under a directory.  This list
is consumed by the tsem_inode_init_security() function to
propagate the inode ownership created in the tsem_create_inode()
function into the tsem_inode structure that was created for the
inode.

The instance_list is used to track the number of temporary files
that a task identity (TASK_ID) has created under a specific
directory.  This list is maintained by the following function:

event.c:_get_task_inode_instance()

This function traverses the list to see if the task identity has
created a temporary file in the directory and if so increments
the temporary file instance count.  If not an entry is created
for the task identity on the list with an instance count of one.

An additional linked list implementation was added to the
tsem_model structure with its own mutex lock.  This list contains
structures that identify inodes that are created in order to
support mount points within the namespace.  This allows the
pathname for the mountpoint to be identified as a temporary
pathname even when it is 'covered' by the inode for the mount.

In the case of an inode for a temporary file the following key
values are output in the context of the JSON path structure:

instance
owner

The instance number is the instance of an inode created under a
specific directory by the task identity that is identified by the
owner key.

The contribution to the CELL identity for the path of a temporary
file is represented by the following:

PATH_ID = HF(DIRPATH || INSTANCE || OWNER)

Where:

	HF: Compression function for the modeling namespace.

	DIRPATH: Directory path containing the temporary file.

	INSTANCE: u64 instance file of the inode that was created.

	OWNER: The task identity of the process creating the file.

This contribution to the CELL identity is implemented in the
map.c:add_temp_file() function.

The generation of the CELL identity was moved from the switch
statement in the map.c:get_cell_mapping() function into the
map.c:add_inode_create() function.

Since a file that is created will have no correct definition for
its digest file the event.c:add_file_digest() function was
updated to check whether the backing inode for a file has been
created.  If so the digest value is set to the zero digest value
for the modeling context.

This implementation is an initial approximation of support for
temporary files and additional enhancements are anticipated.