Use an invariant UUID value for tmpfs filesystems.
It appears that each time a tmpfs filesystem is mount a new and
random UUID is generated.  Since the contribution of an inode to
a CELL identity includes the UUID of the filesystem supporting
the inode, this can cause changes in the state value of a
security model namespace.

This had been addressed, in part, previously by setting the UUID
value a value that was filled with the 0xff byte pattern.

With further experience it is now apparent that the correct
location to address this is when an inode is mapped into the CELL
identity.  The map.c:add_inode() function was modified to use
the NULL UUID value as the uuid for an inode on a tmpfs
filesystem.

The setting of the UUID field in the event.c:get_inode_mknod()
was dropped.