Use size of pathname for pseudonym generation.
Testing indicated that file pseudonym's were not working properly
for internally modeled security namespaces.

This problem was sourced to an artifact of the
model.c:generate_pseudonym() function having previously used the
hash of the pathname rather than the pathname itself for
pseudonym generation.  The call to the crypto_shash_finup()
function was using the digest size of the compression function
being used for the security modeling namespace rather than the
length of the pathname.