Reset inode instance identifiers on namespace termination.
In order to maintain invariance between subsequent invocations of
a workload in a security modeling namespace the task inode
instance identifiers need to be cleared for the security modeling
contex/namespce that created the instance identifiers.  If this
is not done a subsequent invocation of the modeling namespace
will increment the identifiers from the last used identifier
pertubations in the security state of the workload.

In order to address this, a strategy was implemented which tracks
all directory inodes that have subordinate files or directories
created in them in the tsem_context structure created for the
namespace.  The following members were added to the tsem_context
structure in order to support this list:

inode_mutex
inode_list

The tsem_inode_entry structure was introduced to support the
members of this list.

The event.c:register_directory() function was implemented in
order to add an entry to the inode context list.  It is invoked
by the register_inode_create() function as part of the handling
of the TSEM_INODE_CREATE handler.

The namespace.c:wq_put() function that handles the asynchronous
release of a security modeling context was updated to search the
context inode list for any entries prior to the release of the
context.  The _release_inode_instances() function is called for
each directory inode in order to traverse and remove the list of
inode instance identifiers.

The final modification was made to the
tsem.c:tsem_inode_free_security() function to conduct
functionality similar to what is executed when the security
modeling namespace terminates.  This is to handle the release of
instance identifiers for directories that were created and then
removed as part of the normal operation of the namespace.
Failing to do this would result in the leakage of the instance
structures that had been attached to the directory inode for
files created in the directory.