Allow event cache size to be set for root only export.
Previously the size of the event cache for export only root
modeling namespaces was set with a heuristic that was five times
the static event cache size.  This was inflexible and created
issues with booting a Debian implementation with an export only
root model.

The name of the function to setup root only export modeling was
changed from the following:

tsem_ns_export

To the following:

tsem_ns_export_root

In order to clarify the role of the function.

The function was modified to take an unsigned integer as its
argument.  The tsem.c:set_ready() function was modified to call
the setup function with the default magazine size or the value
specified by the command-line directive.