Properly initialize zero digest for root namespace.
Testing indicated that temporary files created in the root
security modeling namespace were getting set to a digest value of
all null bytes.

This was traced to the fact that the digestsize variable in the
tsem.c:configure_root_digest() function was not being
initialized.  This resulted the computed value of the zero digest
not being copied into the zero_digest member of the root security
modeling context.

On review of the function the digestsize variable was dropped and
replaced with a call to the crypto_shash_digestsize() function in
the memcpy() function that initialized the zero digest value in
the root context.