Localize inode instance information to path description.
A review of the code indicated that there was no need to carry
instance and ownership information for inodes attached to
temporary files that are created in the context of a security
modeling namespace in the tsem_inode_cell structure.  This commit
drops the created, creator, instance and owner members from that
structure.

The check for whether or not a dentry represents a temporary file
in the map.c:add_dentry() function was modified to use the
instance information in the tsem_path structure rather than the
now deleted information in the tsem_inode_cell structure.

The functions that were copying instance information from the
tsem_inode structure were modified to drop the copies into the
tsem_inode_cell structure leaving only the copy of information
into the tsem_path structure.