Drop specialized capability bit for TSEM control access.
Previously a new capability bit had been implemented to control
the ability to administrative control a security modeling
namespace.  One of the review comments from Casey Schaufler was
to dispute the need for a specialized capability.  He instead
indicated that CAP_MAC_ADMIN should be used.

Our premise for implementing a new security capability was that
the creation and management of a security modeling namespace was
different than the notion of administrating a mandatory access
control policy.  In addition, there is no reason why a security
modeling namespace would not be implementing a standard mandatory
access control.

Since the security predicate for a security modeling namespace is
to disallow any administrative control over the namespace, there
is a need to drop the security capability for administrating a
namespace.  This thus disallows the ability to administrate
standard mandatory access controls within the security modeling
namespace, that be extension implies that standard mandatory
access controls may have limited utility in a security modeling
namespace.

An example of this model would be to use TSEM to validate the
security attributes used for classic label based policies that
are run within a security modeling namespace.

Given the desire to move this initiative forward the decision was
made to drop the notion of a specialized capability bit in favor
of the use of CAP_MAC_ADMIN.  While this is an inferior
implementation it satisfies the tension between the code not
being accepted and advancing its acceptance.