Update control plane ABI documentation.
To date the Linux reviewers have focused on the TSEM ABI documentation
and to some extent the primary TSEM reference documentation.

This commit updates the ABI documentation to address the issues
that were raised and to make the ABI documentation more
tractable.

The description of the JSON encoding of the events is being
transferred to the TSEM reference documentation.  Attempting to
include that information in the description of either the
trajectory, forensics or export control plane file makes the
document a bit difficult to format and follow.