Eliminate bypass of events involving a pseudo-filesystem.
A long standing scheme has been the bypass of a number of inode
and file related security handlers that involved files that were
on pseudo-filesystem implementations.  This commit eliminates
those bypasses and enables full model resolution.

The tsem.c:bypass_filesystem() function was renamed to
pseudo_filsystem() in order to represent that the intent of the
function is to not bypass a filesystem but return a boolean
indicating whether or not an inode is on a pseudo-filesystem.

The pseudo_file member was added to the 'in' union member of the
tsem_file_args structure.  The pseudo_filesystem() function is
used to the set the value of this member in the tsem_file_open()
function.

The event.c:add_file_digest() function was modified to take a
boolean argument to indicate whether or not the file being acted
on arises from a pseudo-filesystem.  This keeps digest
functionality segregated in this function.