Handle opens for control plane files.
Eliminating bypass of pseudo-filesystem based files resulted in
deadlocks in externally modeled namespaces.  This is a classic
'Heisenberg Deadlock' bug in modeled systems.

This issue was addressed by establing an additional enumeration
type in the tsem_task_trust enumeration.  This new inode type
indicates an inode that is attached to one of the control
Bypass modeling of control plane files.

The removal of bypassing the pseudo-filesystems was a deadlock on
externaly modeled securitiy namespace.  This was traced to a
classic 'Heisdenberg Deadlock' that was caused by trying to model
an event that was required for the modeling of the namespace
to start.

This was fixed by adding a fourth state to the status of an
inode.   This state is assigned to the inodes of the control
plane files including the inode for the file that was created to
surface event descriptions.

A check was added to the tsem.c:tsem_file_open() function to
determine if the file that is the subject of the TSEM_FILE_OPEN
security event is a control plane file.  If this is found to be
true a check against the CAP_MAC_ADMIN capability is done.  If
the process has the capability a zero return value is returned
indicating the open is authorized.  A process lacking the
capability and attempting the open is denied.

The fs.c:_init_inode() function was introduced in order to check
the status of a pseudo-file creation and to set the inode status
if the file had been created.