Standardize on handling of loops locating an entry.
Some of the list traversal loops that were attempting to locate
a specific entry were traversing the entire loop rather than
existing after the target entry was located.

Convert those loop instances to set the located entry and then
issue a 'break' command to exit out of the loop rather than
continueing over the remainder of the loop.