Only register inode creations in modeling namespaces.
Code review indicated that there was no need to register inodes
in the root namespace context as it will never be released.  The
rationale for registering inodes in a namespace context was to
provide a mechanism for cleaning task identity creation instances
when the namespace was released.

Check the security modeling namespace identity and only register
inode creations in subordinate namespaces.