Adjust atomic magazine sizes and cacheing.
Testing with the 6.8 kernel indicated that the system workqueue
priority scheduling had somehow changed and the asynchronous
magaizing replenishment processes were not being scheduled fast
enough to prevent depletion of the structure magazines for
security event handlers that were running in atomic context.

The asynchronous replenishment of the event and export structures
were changed from the system_wq to the system_highpri_wq.  This
allowed the root magazine size to be sufficient for a boot but
occassional failures were noted when the system was being
shutdown.

The root magazine size was increased to 128 that seems to provide
error free allocations.

For internally modeled namespaces a cache value of 16 was
sufficient with the system_highpri_wq but as a matter of safety
the default cache size was increased to 32.

Event with the high priority queue was insufficient to sustain
error free execution of externally modeled domains.  Increasing
the cache size to 128 provided consistent error free executions.