Change security event handler for TASK_ID generation.
Previously the security_bprm_committing_creds security hook was
used to generate the TASK_ID for a new binary that was executed.
This resulted in a situation where the trajectory entry displayed
one set of COE parameters while different COE parameters were
used for the generation of the TASK_ID.

Change to the security_bprm_committed_creds security hook to
trigger the generation of the TASK_ID.  This produces a security
event description that is consistent with the TASK_ID that is
generated.