Do not require attributes for security_inode_init_security.
The integration of the Integrity Measurement Architecture (IMA)
into the LSM infrastructure introduced a conditional check that
denied access to the security_inode_init_security event handler
if the LSM extended attribute 'blob' size is 0.

Modify the function so that it removes the need for a non-zero
extended attribute blob size and bypasses the memory allocation
and freeing that is not needed in that case.

Use a break statement to exit the loop that is iterating over
the defined handlers for this event.  The checks for how to
handle cleanup after the handlers have been called are
centralized at the end of the loop.

Use two exit labels.  One of which handles the case where no
exended attributes have been defined and one that handles the
case where memory allocated for the extended attribute processing
is cleaned up.

Signed-off-by: Greg Wettstein <greg@enjellic.com>