Replace tsem ready check with static key.
The check of the tsem_ready variable is only needed during early
system boot before all of the infrastructure needed by TSEM is
initialized.  Given that the variable is checked in a number of
security hooks and the check is no longer needed this
infrastructure is a candidate for replacement with a static key.

Conver the tsem_ready boolean to a statick key that is named
tsem_not_ready in order to simplify understanding of the code.

Replace the following conditional and action:

if (unlikely(!tsem_ready))
	return 0;

With the following:

if (static_branch_unlikely(tsem_not_ready))
	return 0;

When all of the required infrastructure for TSEM is available
deactivate the branch with the following command:

static_branch_disable(&tsem_not_ready);