Update copyright date and documentation.
This commit updates the copyright date of all the sources files
from 2023 to 2024.  This have been a previous criticism from
kernel reviewers.

Significant expansions to the documentation of each of the source
files was also implemented.  Each file now has a full description
of the overall purpose of the file and how it interacts to
support the entire TSEM implementation.