Use NULL pointer mapping for PR_GET_PDEATHSIG argument.
Invariance testing indicated an occassional variation in the
security state value of the workload being tested.  This was
tracked to the fact that argument 2 to the security_task_prctl
was changing between executions of the workload.

A review of the documentation for the PR_GET_PDEATHSIG option to
the security_task_prctl indicated that this argument is a
pointer to return the signal.  This is obviously an unstable
value between workloads.

Modify the map.c:get_cell_mapping() function to substitute a NULL
value for the argument 2 when the option is PR_GET_PDEATHSIG.