Update TSEM to compatibility with 6.8 kernel.
This commit is the minimum set of changes required to create a
functional implementation for the 6.8 release of the kernel that
will be the target for the V3 patch release.

A change was introduced in the Linux Security Module sub-system
that implements 'stacking' of LSM's.  While this is not a
relevant issue with TSEM there are functional changes required in
order to register TSEM as a valid LSEM.

The most significant is the definition of the id value that is
used to identify the LSM.  The next sequential id of 111 was
used.

This new registration architecture also requires a modification
to the security/security.c file in order to increment the count
of the number of available LSM's.

A lsm_id structure declaration is now needed for each LSM.  In
the case of TSEM this is used to set the name of LSM as well as
its identification number.  The name of the LSM in the
security_add_hooks() function is replaced with a pointer to the
identification structure.

Changes were needed to the signatures of the following event
handlers:

tsem_capget
tsem_bprm_committing_creds
tsem_quotactl

These changes required some concurrent changes in the tsem.h
header file in order to avoid errors.

 struct lsm_blob_sizes tsem_blob_sizes __ro_after_init = {
.lbs_task = sizeof(struct tsem_task),
.lbs_inode = sizeof(struct tsem_inode),
.lbs_ipc = sizeof(struct tsem_ipc),
.lbs_xattr_count = 1

The integration of the LSM/IMA/EVM system in 6.8 resulted in a
addition of the .lbs_xattr_count member to the lsm_blob_sizes
structure.  This is to allow an LSM to reserve space for its own
extended attributes.

Along with this change was a conditional check to the
security_inode_init_security() function that simply returns with
no execution of the handlers if the extended attribute blob size
is zero.

This precludes the use of this handler in situations where
extended attributes are not defined.  This can be caused by
stacking a series of LSM's that do not use extended attributes.

A solution was proposed to the Linux security maintainers to
restore the previous behavior but the response was that LSM's
could simply allocate an extended attribute if they wanted access
to the handler.

TSEM registers a handler for the security_inode_init_security()
function in order to support inodes that are created in the
context of a security modeling namespace.  In order to retain the
functionality of this handler a single extended attribute request
was registered for the TSEM specific 'blobs'.

This commit generates checkpatch warnings secondary to new checks
that have been implemented in the checkpath regarding the
deprecation of various str* functions.  These will be addressed
in subsequent commits in order to isolate the changes.