Initialize inode creation mutex for export only root namespaces.
Testing indicated that booting in tsem_mode=1 resulted in a
kernel panic.  This was traced to the fact that when support for
temporary inode instances was added the initialization of the
context->inode_mutex and context->inode_list had been missed in
the namespace.c:tsem_ns_export_root() function.

Adding these initializations addressed this regression.