Implement export only modeling of root namespace.
This commit provides support for export only of security event
descriptions for the root security modeling namespace.  Selecting
root modeling is done by specifying tsem_mode=3 as a kernel
command-line option.

This update provides the following three modes that can be
specified with the tsem_mode variable, ie:

	0: Full modeling.
	1: No modeling of the root modeling namespace.
	2: Export only of events from the root modeling namespace.

The most significant change needed to support export only
modeling of the root namespace was to add the the
tsem_ns_export() function to the namespace.c file.  This function
allocates all of the infrastructure needed to support exporting
security event descriptions to userspace.  The root security
modeling context is replaced with this newly allocated context.

This function is invoked by the set_ready() function in the
tsem.c file since the TSEM control plane needs to be setup in
order to support the root export modeling.

Security events for the root modeling namespace are surfaced
through the following pseudo-file:

/sys/kernel/security/tsem/ExternalTMA/0

All of the security event descriptions generated are stored until
they are read by a userspace agent.  Failing to read the events
can result in all of system memory being consumed by the linked
list of security event descriptions.