Change lock status defines.
One of the review comments for the V3 release from Casey
Schaufler was a consistency issue with the CPP define that is
used to indicate whether or not the handler is called in atomic
context.

Change the LOCKED define to LOCK so that it pairs conceptually
better with the NOLOCK define.