Improve correlation between event enumerations and descriptions.
One of the review comments from Casey Schaufler in the V3 release
was the fact that having the tsem_event_type enumeration and the
tsem_names array in two separate files.  The suggestion was that
it would not be possible to maintain a correlation between these
two related data structures.

The suggestion had been to have an array of structures that would
use the following member initialization:

{ "string name", TSEM_ENUMERATION_NAME}

The issue with that is the enumerated type is used to select the
element of the array that is to be used.  There would be no
reason to ever use the enumerated type in the structure.  The
effect would be to double the size of the current array with no
benefit.

As an alternative strategy each member of the tsem_names array
had the enumeration type that references that array member
specified as a comment.

Documentation was added to both the tsem_event_type enumeration
and the tsem_names array to indicate that any changes to either
the enumeration constants or the array of names required
coordinated changes between these two entities.