Convert TSEM mode arguments to strings.
In his review Casey Schaufler had suggested that if additional
mode parameters would be considered a switch statement should be
used for handling the numeric values of the tsem_mode argument.
In passing he thought it strange that more descriptive string
values had not been used.

Using a string argument is arguably more clear than using integer
values so this scheme was adopted.

An array of structures was created that pair the string argument
with the enumerated type that describes the mode.  The
tsem.c:set_modeling_mode() function traverses this array and sets
the mode argument if a recognized option was found.

If an invalid argument was specified the system defaults to full
modeling.

The previous argument of '1', used to specify no root modeling is
now specified by the following keyword:

no_root_modeling

The previous argument of '2', used to specify that the root
modeling namespace operate only in export mode is now specified
by the following keyword:

root_export_only