Simplify namespace freeing logic.
In his review Casey Schaufler noted that by re-arranging the
logic in the tsem.c:tsem_task_free() function a line of code
could be saved.

Change the logic for releasing the security modeling namespace
from the following:

if (!ctx->id)
	return

To:

if (ctx->)
	tsem_ns_put(ctx)

With execution then allowed to hit the end of the function.

The return statement was dropped since this is a void returning
function.