Remove warning on lack of inode attribution.
Kernel warnings were noted when root namespace modeling was
disabled.  This was secondary to the fact that inodes that were
created when modeling in the root security namespace was disabled
did not have an entry attached to the containing inode.

Retaining this warning would have required additional logic to
determine if an inode description should be expected.  In
addition, since kernel warnings are now generally frowned upon,
this check was removed.