Implement selectable event handling.
This commit begins the process of preparing TSEM to handle
multiple security models through the use of loadable modules.

The first step in the process is to provide a framework for
implementing the ability for different models to select which
security events they will handle.

The tsem_context_ops structure was added to provide a methods
container for functions and data which implement the desired
model.  This structure is included in the tsem_context structure
as it defines the characteristics of the security modeling
namespace.

The initial data in the tsem_context_ops structure is an array of
booleans that is equal in size to the number of security event
handlers that are implemented.  Each member of the array
specifies whether or not a particular security event is bypassed
with no action.  The array, like the tsem_names array and the
actions array, is indexed via the tsem_event_type enumeration.

The model0.c compilation unit was introduced to contain the
declaration of the tsem_context_ops structure for the default
deterministic TMA that TSEM implements.  This compilation unit
also contains the bypass array that contains no bypass
declarations since the default TMA implements handlers for all
security events.

The namespace.c:tsem_ns_create() function was modified to take a
pointer to a tsem_context_ops structure as an argument to the
function.  In the initial implementation the call sites for this
function in the fs.c file simply passes a pointer to the
tsem_model0_ops structure as no support is initially available
for multiple models.

The tsem.c:bypass_event() function provides the common point for
interrogating the model specific event bypass array.  This
required changing the signature of the function to take a
tsem_event_type enumeration argument that is used to index the
array to determine the bypass status for the event.  This
required that each call to the function was modified to take the
enumeration description of the event handler from which it is
being called.