Implement framework for registering TSEM models.
In order to support the ability to implement alternate TSEM
security models with loadable modules the nsmgr.c file was
introduced.  This file 'manages' the various security models that
are available to TSEM.

In order to support this management the .name member was added to
the the tsem_context_ops structure to allow a name to be assigned
to the security model.  Models are located, registered and
released based on this name.

The following two functions that are visible to loadable modules
are supplied for the registration and release of security models:

tsem_nsmgr_register

tsem_nsmgr_release

A module registers a security model by providing a pointer to a
tsem_context_ops structure that defines the operational
characteristics of the model.  A pointer to module structure is
also provided so that each security modeling namespace that uses
a model instance can hold a reference to the module to prevent
its release.