Implement support for selecting a loaded security model.
The TSEM control plane was modified so that a write to the
control file to create a security namespace now accepts the
following argument:

model=NAME

Where NAME is the name of a model implemented by a kernel
loadable module.

If no model is selected the 'model0' model is automatically
selected.  If a model with the specified name cannot be found on
the model list the creation of the namespace fails.

The following external functions were added to the nsmgr.c file
in order to support obtaining and releasing a reference to a
loaded security model:

tsem_nsmgr_get()

tsem_nsmgr_put()

The tsem_nsmgr_get() function takes an argument to the security
model that a reference is to be obtained from.  If found the
function returns a pointer to the operations structure for the
model.

The tsem_nsmgr_put() function takes a pointer to a model
operations structure and releases the reference to the module
that implements the model.

This commit makes functional improvements to the
nsmgr.c:tsem_nsmgr_register() function including the actual
additional of a new entry to the list that holds the current
active models.