Make security module registration dependent on module support.
Since alternate security models can only be registered in the
kernel by loadable module there is no need to have the model
manager implemented if kernel modules are not enabled.

Make the inclusion of the nsmgr.o object file dependent on
CONFIG_MODULE.

The globally visible interfaces for module registration were
moved to the nsmgr.h file so that alternate stub functions can be
generated provide functional interfaces in the absence of
loadable modules.