Support demand loading of security models.
With basic module based security models working it was a straight
forward exercise to implement demand loading of modules when a
particular security model was requested.

The nsmgr.c:tsem_nsmgr_get() function was modified to attempt to
find a pointer to a security model in an isolated critical
region.

If the lookup returns a NULL pointer indicating the model is not
available a request_module() call using the name of the model is
made.  If a subsequent lookup of the model still fails a failure
is returned to the caller.

If the initial lookup, or the lookup after the module request
succeeds, a reference is taken to the module and the the model
pointer is returned to the caller.