Return event characterization code from model0 implementation.
The bulk of the code that converts the arguments to the security
event handlers into their retained versions had been placed in
the code that contains the declaration for the default, model0,
TSEM security modeling implementation.

This architecture proved problematic at a number of levels so the
event characterization code was returned to the event.c file.

The function that conducts the initialization was renamed from
the following function name:

tsem_model0_event_init()

To the following:

tsem_event_generate()

In the interest of a common approach the following of the
tsem_context_ops structure:

event_init

Was renamed to the following:

init