Handle untrusted tasks immediately.
In the tsem.c:dispatch_event() function the interception of an
untrusted event was being done after the TSEM event description
structure was generated.  The initialization of the structure is
unneeded since the handling of an untrusted event is not
dependent on anything othen than the event type description.

Move the test for an untrusted function and its handling to be
the first task that is handled in the dispatch_event() function.