Initialize event structure reference on allocation.
Further review of the tsem.c:dispatch_event() function suggested
a need to change the reference architecture for the TSEM event
description structures.

The reference to the structure was previously initialized in the
event.c:tsem_event_init() function.  In the case of an untrusted
event the event description is not initialized which means that
an external function would need to be supplied to conduct the
freeing of the event structure.

In order to increase the reliability of the management of the
event structure the initialization of the reference was moted to
the event.c:tsem_event_init() function.  At this point the
tsem_event_put() function can always be used to release the
structure and return it to the event cache.

The tsem.c:dispatch_event() function now uses a single exit point
that releases the reference to the structure that was taken by
its initialization.  If the event is being exported an additional
reference will be taken that will then be removed when the event
description is read.