Update generation of COE mapping.
The map.c:get_COE_mapping() function was using open coded
additions of the characteristics of the context of execution
rather than the helper functions used to add fixed size integers.

Move the get_COE_mapping() function after the utility functions
and update the function to use the utility functions to generate
the mapping value for the context of execution.