Use only a single hash update for the creation of the CELL value.
Previously the map.c:get_cell_mapping() function computed the
CELL identity with calls to the finalization of the cryptographic
digest value in each of the clauses of the switch statement that
directs the generation of the identity based on the event type.

This resulted in a significant amount of unneeded code in each
clause of the switch statement that can be eliminated by
conducting the digest finalization after all the digest input has
been generated.  Move the digest finalization to the end of the
function and gate the finalization on whether or not any errors
have been encountered in the generation of the digest input.