Update event initialization handling.
Testing with the tsem-integrity loadable model demonstrated a
long standing issue with how event characteristics are generated.

The first change was to check the event event and if the event
was for the TSEM_BPRM_COMMITTED_CREDS event the
tsem_event_generate() function is called without calling the
model specific initialization code.  This was done so that the
tsem_map_task() function would always be conducted in a
standardized fashion given its importance to the TSEM security
model.

The other important issue addressed was to remove the call the
kmem_cache_free() function that was being invoked if an error
occurred in event generation or if the model specific
initialization function indicated the event should be bypassed.

This is to satisfy the contract with the event handling
infrastructure which requires that once the tsem_event_allocate()
function is successfully called the event description structure
should only be retired using a tsem_event_put() call.