Implement support for locking TSEM modeling state.
One of the potential security issues with TSEM models implemented
with loadable modules is the possibility of removing a model or
installing an alternate model and using that.  The commit
addresses these potential vulnerabilities by implementing support
for locking the registration state of models.

Writes to the TSEM control file were modified to accept the
following keyword:

lock

Which is the command signal to irreversibly lock the model state.

When issued by userspace the locked variable in the nsmgr.c file
is set to a value of true, there is no API implemented to reverse
this value.

In addition any model modules that are loaded have their
reference count increased by one so that the modules cannot be
unloaded.

All of the functionality for locking the model state is
implemented in the nsmgr.c:tsem_nsmgr_lock() function.

In order to close any potential races the following kernel
command-line parameter was also implemented:

tsem_lock

Which is called by the kernel setup code and sets the following
variable in the nsmgr.c file to true:

setup_locked

This variable is annotated in order to have it set to read-only
after system initialization so locking the model state is 'more'
irreversible than the lock implemented by userspace, presumably
after models are loaded.