Update documentation for the 'model' control argument.
The original documentation for the model control argument was
incorrect with respect to specifying that the security model name
could be set by the THIS_MODULE CPP define.  Update the
documentation so that the relevant CPP define is as follows:

KBUILD_MODNAME

Also clarify that the security model name does not need to match
the name of the module as displayed by the lsmod command.