Implement event numbering.
Testing and data analysis indicated that it would be useful to
have the events generated sequentially as they are generated in
the context of a security modeling namespace.  This commit
implements that numbering.

The JSON event struct was updated to include the following key:

number

Whose value is a 64-bit integer value reflecting the sequence
number of the event in the security modeling namespace that
generates the event.

The tsem_context structure was updated to include a 64 bit value
that represents the current event number.  Since this event
number is a characteristic of the event that may be retained a
similar value was created in the tsem_event structure to preserve
the value for the lifetime of the event.