Export security namespace identifier.
Work on modeling of the data suggested that it would be helpful
to have an identification of the security modeling
namespace/context that generated the event.

The security modeling context is exported in the event{}
structure with the following key:

context

The value of the key is the 64-bit security context identifier.

A context member was added to the tsem_event structure so that
the security modeling namespace identifier can be captured for
the event. The value is set in the event.c:tsem_event_init()
function from the id value of the context structure for the task
calling the event.