Convert namespace timestamp reference to creation of namespace.
Work with data analysis on exported namespaces suggested that a
more consistent approach to the event timestamps would be to
reference the timestamp as relative to the creation of the
namespace rather than the system being booted.

This commit converts the exported timestamps to this reference
system.

A 64 bit integer timestamp member was added to the tsem_context
structure in order to capture the number of nanoseconds since
the boot of the system in the modeling namespace context.  This
value then serves as the new reference value for event
timestamps.  The reference time for the security modeling
namespace is taken when the tsem_ns_create() function is called
to create the new subordinate modeling namespace.

The tsem_event_init() function was updated to capture the
relevant timestamp value when the tsem_event structure is
initialized for the event.