Implement customized event characterization.
In order to increase the flexbility of models that can be
implemented as loadable modules this commit provides a framework
for allowing the characterization of a security event to be
customized.

The following member was added to the tsem_context_ops structure
in order to specify the function that will be used to initialize
the retained characteristics for the event:

event_init

The majority of the event initialization was transferred from the
event.c to the model0.c file.   The following function is exposed
from this file so that the default parameter initialization can
be used by modules:

tsem_model0_event_init

Since the event structure can be freed by a process that is not
running in the security modeling namespace in which it was
initialized, ie. a workqueue, the event_free member was added to
the tsem_event structure.  The expectation is that the event
initialization code will place a pointer to the appropriate
freeing function during the event initialization.

The tsem_event_free() function was modified to call the event
description freeing structure if the event_free member of the
tsem_event structure is non-NULL.

The return value of the tsem_event_init() function was modified
to have three different possible return values:

< 0: Errors are returned as negatively encoded error values.

= 0: Event modeling is to be bypassed.

> 0: Event is to be modeled.