Implement configurable coefficient mapping policies.
This commit follows the framework that was introduced for the
characterization of security events by implementing the ability
to customize the mapping of the events into a security state
coefficient.

The following member was added to the tsem_context_ops structure
to hold a pointer to the function that conducts the security
state coefficient mapping:

map

In the model.c:tsem_model_event() function the current call to
the mapping function:

retn = tsem_map_event(ep)

Was replaced with the following:

retn = ctx->ops->map(ep)

This commit also corrects the declaration of the tsem_map_event()
function in the tsem.h header file.