Fix handling of inode registration.
When an inode is created in a security modeling namespace the
inode is registered on linked list until the TSEM inode instance
can be initialized by the tsem_inode_init_security() function
when the inode is physically instantiated.

A regression was noted with externally modeled namespaces that
were not recognizing that an inode was being created in the
context of a security modeling namespace.  This was traced to the
fact that the pathname of the inode was being corrupted and thus
not detected by the code in the tsem_inode_init_security()
function.

The reason for the corruption was that the pointer to the inode
pathname that was being used was created by the event
initialization code.  The lifetime of that code was only until
the event was exported, after which the memory was freed and
the pointer was thus not pointing to a valid pathname.

This situation was corrected by copying the pathname into a
buffer allocated with the __getname() function with a pointer to
the buffer being stored in the tsem_inode instance structure.

After detecting the creation of the inode both the pointer to the
pathname and the pointer to the tsem_inode_instance structure
were freed.

This fix also corrects a memory link caused by the previous
implementation not freeing the pointer to the tsem_inode_instance
structure after the structure was taken off the inode
registration list.