Move namespace mount list to namespace context.
The list of inodes that serve as mountpoints was originally
implemented as part of the namespace model description.  This was
a function of the fact that the inode instance implementation was
done with internally modeled namespaces.

Unfortunately this resulted in the instance information not being
included in the security event description that was being
exported to an external trust orchestrator.

Moving the mount list description to the namespace context
structure causes the list to be available to both internal and
externally modeled domains.

Following the same concept as the implementation of the inode
instance list, the buffer for the inode pathname is allocated
rather then using the pathname from the event description
structure.  This prevents the corruption of the pathname when the
event description structure is released after the model response
is generated.