Convert mount list to use an allocated pathname.
The entries on a security modeling namespace mountlist last for
the entire lifetime of the namespace, which could be substantial.
As a result it is wasteful to use a 4K page of memory for a
pathname, given that the vast majority of pathnames will only be
a fraction of a page in size.

Convert the use of __getname and __putname to a manually
allocated character buffer using kstrdup and kfree.